% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_data.R
\name{acf_vec}
\alias{acf_vec}
\title{Estimate autocorrelations of a numeric vector}
\usage{
acf_vec(x, lag_max = 24, ...)
}
\arguments{
\item{x}{Numeric vector.}

\item{lag_max}{Integer. Maximum lag for which the autocorrelation is
estimated.}

\item{...}{Further arguments passed to \code{stats::acf()}.}
}
\value{
A numeric vector containing the sample autocorrelations for lags
\code{1} to \code{lag_max}.
}
\description{
Estimate the sample autocorrelation function of a numeric vector.
}
\details{
\code{acf_vec()} is a small wrapper around \code{stats::acf()}. It returns
the sample autocorrelations as a numeric vector and removes lag 0 from the
output, because lag 0 is always equal to 1 and is usually not needed for
diagnostics.
}
\examples{
library(dplyr)

x <- M4_monthly_data |>
  filter(series == first(series)) |>
  pull(value)

acf_vec(
  x = x,
  lag_max = 12
)
}
\seealso{
Other data analysis: 
\code{\link{estimate_acf}()},
\code{\link{estimate_kurtosis}()},
\code{\link{estimate_mode}()},
\code{\link{estimate_pacf}()},
\code{\link{estimate_skewness}()},
\code{\link{pacf_vec}()},
\code{\link{summarise_data}()},
\code{\link{summarise_split}()},
\code{\link{summarise_stats}()}
}
\concept{data analysis}
