% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{approx_vec}
\alias{approx_vec}
\title{Linear interpolation of NAs.}
\usage{
approx_vec(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\value{
x Numeric vector.
}
\description{
\code{approx_vec} linearly interpolates NAs
   within a numeric vector.
}
