% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_data.R
\name{bind_data}
\alias{bind_data}
\title{Bind data and forecasts row-wise into one tsibble.}
\usage{
bind_data(data, fcsts)
}
\arguments{
\item{data}{A tsibble containing the data provided by \code{split_data(...)}.}

\item{fcsts}{A fable ('forecast table') provided by \code{forecast_model(...)}.}
}
\value{
data A tsibble containing the input data and the forecasts.
}
\description{
Bind data and forecasts row-wise into one tsibble. The key structure of both objects is updated.
}
