% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{clean_data}
\alias{clean_data}
\title{Check, convert and shape the input data.}
\usage{
clean_data(data)
}
\arguments{
\item{data}{A valid tsibble, either in long or in wide format.}
}
\value{
data A tsibble in long format with the following columns:
   \itemize{
      \item{\code{date_time}: The index variable of the tsibble.}
      \item{\code{variable}: Character value. The key variable of the tsibble.}
      \item{\code{value}: Numeric value. The measurement variable of the tsibble.}
      \item{\code{type}: Character value. The type of the observation (i.e. actual, fcst, error, pct_error).}
      \item{\code{target}: Logical value, indicating whether the variable is a target variable or not.}
      }
}
\description{
This function checks whether the input data are a valid tsibble or not (regular spaced in time
   and ordered). Furthermore, implicit missing values are turned into explicit missing values (existing
   missing values are left untouched). If the data are provided in wide format, they are gathered into
   long format. For convenience, the index variable is renamed to \code{date_time}, the key variable is
   renamed to \code{variable} and the measurement variable is renamed to \code{value} and the columns
   \code{type} and \code{target} are added.
}
