% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_errors.R
\name{create_errors}
\alias{create_errors}
\title{Calculate errors and percentage errors.}
\usage{
create_errors(data)
}
\arguments{
\item{data}{A tsibble in long format with standard columns (date_time, variable, etc.).}
}
\value{
errors A tsibble with the following columns:
   \itemize{
      \item{date_time: Index variable}
      \item{variable: Target variable}
      \item{sample: Train or test}
      \item{slice: Time slice}
      \item{horizon: Forecast horizon (n-step ahead)}
      \item{type: Error or percentage error}
      \item{model: Forecasting model}
      \item{value: Measurement variable}
      }
}
\description{
This function calculates the errors (actual - fcst) and percentage errors ((actual - fcst) / fcst * 100).
}
