% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_grid.R
\name{create_grid}
\alias{create_grid}
\title{Create a grid of unique combinations}
\usage{
create_grid(data)
}
\arguments{
\item{data}{\code{train} or \code{test} data from \code{split_data()}.}
}
\value{
A tibble containing the unique combinations.
}
\description{
The function creates a grid of unique combinations for
  key variables and splits.
}
