% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_name}
\alias{create_name}
\title{Create a name for a folder or file}
\usage{
create_name(primary, extension = NULL, add_time = TRUE)
}
\arguments{
\item{primary}{Character value. Primary file name to be used.}

\item{extension}{Character value. File name extension to be used
(e.g. \code{".csv"}, \code{".rda"}).}

\item{add_time}{Logical value. If \code{TRUE}, a time stamp is added in
front of the file name.}
}
\value{
Character value.
}
\description{
The function creates a name for a folder or file.
}
