% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{elec_price}
\alias{elec_price}
\title{Hourly day-ahead electricity spot prices}
\format{
A time series object of class \code{tibble} with 140.352 rows and 5 columns:
   \itemize{
      \item{\code{time}: Date and time index}
      \item{\code{item}: Time series name}
      \item{\code{unit}: Measured unit}
      \item{\code{bidding_zone}: Bidding zone}
      \item{\code{value}: Measurement variable}
      }
}
\source{
\href{https://transparency.entsoe.eu/}{ENTSO-E Transparency Platform}
}
\usage{
data(elec_price)
}
\description{
Hourly tibble with day-ahead electricity spot prices from
  the ENTSO-E Transparency Platform. The data set contains time series data
  from 2019-01-01 00:00:00 to 2020-12-31 23:00:00 for 8 bidding zones
  within Europe (DE, DK1, ES, FI, FR, NL, NO1, SE1).
}
\examples{
data(elec_price)
}
\keyword{datasets}
