% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELM.R
\name{ELM}
\alias{ELM}
\title{Automatic training of ELMs.}
\usage{
ELM(formula, ...)
}
\arguments{
\item{formula}{Model specification (see "Specials" section, currently not in use...)}

\item{...}{Further arguments passed to \code{nnfor::elm()}.}
}
\value{
elm_model An object of class \code{ELM}.
}
\description{
Automatic training of ELMs.
}
