% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_metrics.R
\name{error_metrics}
\alias{error_metrics}
\title{Create forecast accuracy measures.}
\usage{
error_metrics(errors, dim = c("slice", "horizon"))
}
\arguments{
\item{errors}{A tsibble in long format with standard columns (.date, .variable, .value, etc.) provided by the function \code{create_errors(...)}.}

\item{dim}{Character value defining the dimension for the evaluation. Possible values are "slice" or "horizon" or both.}
}
\value{
metrics A tibble with the following columns:
   \itemize{
      \item{variable: Target variable}
      \item{model: Forecasting model}
      \item{dim: Forecast accuracy by slice or by horizon}
      \item{num: The number of the slice or the nth-step forecast}
      \item{type: Training or testing}
      \item{metric: The accuracy metric}
      \item{value: Measurement variable}
      }
}
\description{
This function creates several forecast accuracy measures to evaluate point forecast. By default,
   the following accuracy measures are provided:
   \itemize{
      \item{ME: Mean Error}
      \item{MAE: Mean Absolute Error}
      \item{MSE: Mean Squared Error}
      \item{RMSE: Root Mean Squared Error}
      \item{MPE: Mean Percentage Error}
      \item{MAPE: Mean Absolute Percentage Error}
      }
}
