% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{errors}
\alias{errors}
\title{Calculate forecast errors}
\usage{
errors(fcst, test)
}
\arguments{
\item{fcst}{A \code{fable} containing the forecasts for the models, splits, etc.}

\item{test}{A \code{tsibble} containing the testing data.}
}
\value{
errors A \code{tsibble}.
}
\description{
\code{error} calculates the forecast errors (actual - fcst).
}
