% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EXPERT.R
\name{EXPERT}
\alias{EXPERT}
\title{Automatic train a EXPERT model}
\usage{
EXPERT(formula, ...)
}
\arguments{
\item{formula}{Model specification (see "Specials" section, currently not in use ...)}

\item{...}{Further arguments (currently not in use).}
}
\value{
expert_model An object of class \code{EXPERT}.
}
\description{
Automatic train the EXPERT model.
}
