% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_NAs.R
\name{fill_NAs}
\alias{fill_NAs}
\title{Interpolate missing values.}
\usage{
fill_NAs(data, method = "arima")
}
\arguments{
\item{data}{A tsibble in long format with standard columns (.date, .variable, .value, etc.) provided by the function \code{shape_data(...)}.}

\item{method}{Character value. Either \code{arima} or \code{tslm}.}
}
\value{
data A tsibble in long format with standard columns (.date, .variable, .value, etc.)
}
\description{
This function interpolates missing values. Leading NAs or backcasted, trailing NAs are forecasted
   and centered NAs are interpolated. The user can choose between several modeling procedures.
}
