% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELM.R
\name{fitted.ELM}
\alias{fitted.ELM}
\title{Extract fitted values from a ELM}
\usage{
\method{fitted}{ELM}(object, ...)
}
\arguments{
\item{object}{The time series model used to produce the forecast.}

\item{...}{Further arguments.}
}
\value{
Fitted values as \code{tsibble}.
}
\description{
Extract fitted values from a ELM.
}
