% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EXPERT.R
\name{fitted.EXPERT}
\alias{fitted.EXPERT}
\title{Extract fitted values from a trained EXPERT model}
\usage{
\method{fitted}{EXPERT}(object, ...)
}
\arguments{
\item{object}{An object of class \code{EXPERT}.}

\item{...}{Currently not in use.}
}
\value{
Fitted values as tsibble.
}
\description{
Extract fitted values from a trained EXPERT model.
}
