% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLP.R
\name{fitted.MLP}
\alias{fitted.MLP}
\title{Extract fitted values from a trained MLP model}
\usage{
\method{fitted}{MLP}(object, ...)
}
\arguments{
\item{object}{An object of class \code{MLP}.}

\item{...}{Currently not in use.}
}
\value{
Fitted values as tsibble.
}
\description{
Extract fitted values from a trained MLP model.
}
