% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMEAN.R
\name{fitted.SMEAN}
\alias{fitted.SMEAN}
\title{Extract fitted values from a trained seasonal mean model}
\usage{
\method{fitted}{SMEAN}(object, ...)
}
\arguments{
\item{object}{An object of class \code{SMEAN}.}

\item{...}{Currently not in use.}
}
\value{
Fitted values as tsibble.
}
\description{
Extract fitted values from a trained seasonal mean model.
}
