% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-fable.R
\name{fitted.SMEDIAN}
\alias{fitted.SMEDIAN}
\title{Extract fitted values from a seasonal median model}
\usage{
\method{fitted}{SMEDIAN}(object, ...)
}
\arguments{
\item{object}{A fitted \code{SMEDIAN} model object.}

\item{...}{Additional arguments. Currently not used.}
}
\value{
Fitted values.
}
\description{
Extract fitted values from a fitted \code{SMEDIAN} model.
}
\examples{
library(dplyr)
library(tsibble)
library(fabletools)

train_frame <- elec_price |>
  filter(bidding_zone == "DE") |>
  slice_head(n = 24 * 21) |>
  as_tsibble(index = time)

model_frame <- train_frame |>
  model("SMEDIAN" = SMEDIAN(value ~ lag("week")))

fitted(model_frame)
}
\seealso{
Other SMEDIAN: 
\code{\link{SMEDIAN}()},
\code{\link{forecast.SMEDIAN}()},
\code{\link{model_sum.SMEDIAN}()},
\code{\link{residuals.SMEDIAN}()}
}
\concept{SMEDIAN}
