% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-fable.R
\name{fitted.TBATS}
\alias{fitted.TBATS}
\title{Extract fitted values from a TBATS model}
\usage{
\method{fitted}{TBATS}(object, ...)
}
\arguments{
\item{object}{A fitted \code{TBATS} model object.}

\item{...}{Additional arguments. Currently not used.}
}
\value{
Fitted values.
}
\description{
Extract fitted values from a fitted \code{TBATS} model.
}
\details{
This method is used by \code{fitted()} when extracting fitted values from a
mable containing a \code{TBATS} model.
}
\examples{
\donttest{
library(dplyr)
library(tsibble)
library(fabletools)

train_frame <- elec_price |>
  filter(bidding_zone == "DE") |>
  slice_head(n = 24 * 21) |>
  as_tsibble(index = time)

model_frame <- train_frame |>
  model("TBATS" = TBATS(value, periods = c(24, 168)))

fitted(model_frame)
}
}
\seealso{
Other TBATS: 
\code{\link{TBATS}()},
\code{\link{forecast.TBATS}()},
\code{\link{model_sum.TBATS}()},
\code{\link{residuals.TBATS}()}
}
\concept{TBATS}
