% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/THETA.R
\name{fitted.THETA}
\alias{fitted.THETA}
\title{Extract fitted values from a THETA.}
\usage{
\method{fitted}{THETA}(object, ...)
}
\arguments{
\item{object}{The time series model used to produce the forecast.}

\item{...}{Further arguments.}
}
\value{

}
\description{
Extract fitted values from a THETA.
}
