% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-fable.R
\name{forecast.DSHW}
\alias{forecast.DSHW}
\title{Forecast a DSHW model}
\usage{
\method{forecast}{DSHW}(object, new_data, specials = NULL, ...)
}
\arguments{
\item{object}{A fitted \code{DSHW} model object.}

\item{new_data}{A \code{tsibble} containing future time points.}

\item{specials}{Parsed specials. Currently not used.}

\item{...}{Additional arguments. Currently not used.}
}
\value{
A vector of forecast distributions.
}
\description{
Forecast a fitted \code{DSHW} model.
}
\examples{
\donttest{
library(dplyr)
library(tsibble)
library(fabletools)

train_frame <- elec_price |>
  filter(bidding_zone == "DE") |>
  slice_head(n = 24 * 28) |>
  as_tsibble(index = time)

model_frame <- train_frame |>
  model("DSHW" = DSHW(value, periods = c(24, 168)))

forecast(model_frame, h = 24)
}
}
\seealso{
Other DSHW: 
\code{\link{DSHW}()},
\code{\link{fitted.DSHW}()},
\code{\link{model_sum.DSHW}()},
\code{\link{residuals.DSHW}()}
}
\concept{DSHW}
