% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EXPERT.R
\name{forecast.EXPERT}
\alias{forecast.EXPERT}
\title{Forecast a trained EXPERT model}
\usage{
\method{forecast}{EXPERT}(object, new_data, specials = NULL, ...)
}
\arguments{
\item{object}{An object of class \code{EXPERT}.}

\item{new_data}{Forecast horizon (n-step ahead forecast)}

\item{specials}{Specials are currently not in use.}

\item{...}{Additional arguments for forecast method.}
}
\value{
An object of class \code{fable}.
}
\description{
Forecast a trained EXPERT model.
}
