% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-fable.R
\name{forecast.MEDIAN}
\alias{forecast.MEDIAN}
\title{Forecast a median model}
\usage{
\method{forecast}{MEDIAN}(object, new_data, specials = NULL, ...)
}
\arguments{
\item{object}{A fitted \code{MEDIAN} model object.}

\item{new_data}{A \code{tsibble} containing future time points.}

\item{specials}{Parsed specials. Currently not used.}

\item{...}{Additional arguments. Currently not used.}
}
\value{
A vector of forecast distributions.
}
\description{
Forecast a fitted \code{MEDIAN} model.
}
\examples{
library(dplyr)
library(tsibble)
library(fabletools)

train_frame <- M4_monthly_data |>
  filter(series == first(series)) |>
  as_tsibble(index = index)

model_frame <- train_frame |>
  model("MEDIAN" = MEDIAN(value ~ window()))

forecast(model_frame, h = 12)
}
\seealso{
Other MEDIAN: 
\code{\link{MEDIAN}()},
\code{\link{fitted.MEDIAN}()},
\code{\link{model_sum.MEDIAN}()},
\code{\link{residuals.MEDIAN}()}
}
\concept{MEDIAN}
