% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLP.R
\name{forecast.MLP}
\alias{forecast.MLP}
\title{Forecast a trained MLP model}
\usage{
\method{forecast}{MLP}(object, new_data, specials = NULL, ...)
}
\arguments{
\item{object}{An object of class \code{MLP}.}

\item{new_data}{Forecast horizon (n-step ahead forecast)}

\item{specials}{Specials are currently not in use.}

\item{...}{Additional arguments for forecast method.}
}
\value{
An object of class \code{fable}.
}
\description{
Forecast a trained MLP model.
}
