% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-fable.R
\name{forecast.NAIVE2}
\alias{forecast.NAIVE2}
\title{Forecast a Naive2 model}
\usage{
\method{forecast}{NAIVE2}(object, new_data, specials = NULL, ...)
}
\arguments{
\item{object}{A fitted \code{NAIVE2} model object.}

\item{new_data}{A \code{tsibble} containing future time points.}

\item{specials}{Parsed specials. Currently not used.}

\item{...}{Additional arguments. Currently not used.}
}
\value{
A vector of forecast distributions.
}
\description{
Forecast a fitted \code{NAIVE2} model.
}
\details{
Point forecasts are calculated using \code{naive2()}. Forecast
distributions use a normal approximation based on a random walk applied
to the seasonally adjusted response. These forecast distributions are not
part of the original M4 Naive2 specification.
}
\examples{
library(dplyr)
library(tsibble)
library(fabletools)

train_frame <- M4_monthly_data |>
  filter(series == first(series)) |>
  as_tsibble(index = index)

model_frame <- train_frame |>
  model("NAIVE2" = NAIVE2(value ~ season(12)))

forecast(model_frame, h = 18)
}
\seealso{
Other NAIVE2: 
\code{\link{NAIVE2}()},
\code{\link{fitted.NAIVE2}()},
\code{\link{model_sum.NAIVE2}()},
\code{\link{residuals.NAIVE2}()}
}
\concept{NAIVE2}
