% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-fable.R
\name{forecast.SMEAN}
\alias{forecast.SMEAN}
\title{Forecast a seasonal mean model}
\usage{
\method{forecast}{SMEAN}(object, new_data, specials = NULL, ...)
}
\arguments{
\item{object}{A fitted \code{SMEAN} model object.}

\item{new_data}{A \code{tsibble} containing future time points.}

\item{specials}{Parsed specials. Currently not used.}

\item{...}{Additional arguments. Currently not used.}
}
\value{
A vector of forecast distributions.
}
\description{
Forecast a fitted \code{SMEAN} model.
}
\examples{
library(dplyr)
library(tsibble)
library(fabletools)

train_frame <- M4_monthly_data |>
  filter(series == first(series)) |>
  as_tsibble(index = index)

model_frame <- train_frame |>
  model("SMEAN" = SMEAN(value ~ lag("year")))

forecast(model_frame, h = 12)
}
\seealso{
Other SMEAN: 
\code{\link{SMEAN}()},
\code{\link{fitted.SMEAN}()},
\code{\link{model_sum.SMEAN}()},
\code{\link{residuals.SMEAN}()}
}
\concept{SMEAN}
