% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-fable.R
\name{forecast.SMEDIAN}
\alias{forecast.SMEDIAN}
\title{Forecast a seasonal median model}
\usage{
\method{forecast}{SMEDIAN}(object, new_data, specials = NULL, ...)
}
\arguments{
\item{object}{A fitted \code{SMEDIAN} model object.}

\item{new_data}{A \code{tsibble} containing future time points.}

\item{specials}{Parsed specials. Currently not used.}

\item{...}{Additional arguments. Currently not used.}
}
\value{
A vector of forecast distributions.
}
\description{
Forecast a fitted \code{SMEDIAN} model.
}
\examples{
library(dplyr)
library(tsibble)
library(fabletools)

train_frame <- elec_price |>
  filter(bidding_zone == "DE") |>
  slice_head(n = 24 * 21) |>
  as_tsibble(index = time)

model_frame <- train_frame |>
  model("SMEDIAN" = SMEDIAN(value ~ lag("week")))

forecast(model_frame, h = 24)
}
\seealso{
Other SMEDIAN: 
\code{\link{SMEDIAN}()},
\code{\link{fitted.SMEDIAN}()},
\code{\link{model_sum.SMEDIAN}()},
\code{\link{residuals.SMEDIAN}()}
}
\concept{SMEDIAN}
