% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-fable.R
\name{forecast.SNAIVE2}
\alias{forecast.SNAIVE2}
\title{Forecast a SNAIVE2 model}
\usage{
\method{forecast}{SNAIVE2}(object, new_data, specials = NULL, ...)
}
\arguments{
\item{object}{A fitted \code{SNAIVE2} model object.}

\item{new_data}{A \code{tsibble} containing future time points.}

\item{specials}{Parsed specials. Currently not used.}

\item{...}{Additional arguments. Currently not used.}
}
\value{
A vector of forecast distributions.
}
\description{
Forecast a fitted \code{SNAIVE2} model.
}
\examples{
library(dplyr)
library(tsibble)
library(fabletools)

train_frame <- elec_price |>
  filter(bidding_zone == "DE") |>
  slice_head(n = 24 * 21) |>
  as_tsibble(index = time)

model_frame <- train_frame |>
  model("SNAIVE2" = SNAIVE2(value))

forecast(model_frame, h = 24)
}
\seealso{
Other SNAIVE2: 
\code{\link{SNAIVE2}()},
\code{\link{fitted.SNAIVE2}()},
\code{\link{model_sum.SNAIVE2}()},
\code{\link{residuals.SNAIVE2}()}
}
\concept{SNAIVE2}
