% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-fable.R
\name{forecast.TBATS}
\alias{forecast.TBATS}
\title{Forecast a TBATS model}
\usage{
\method{forecast}{TBATS}(object, new_data, specials = NULL, ...)
}
\arguments{
\item{object}{A fitted \code{TBATS} model object.}

\item{new_data}{A \code{tsibble} containing future time points.}

\item{specials}{Parsed specials. Currently not used.}

\item{...}{Additional arguments. Currently not used.}
}
\value{
A vector of forecast distributions.
}
\description{
Forecast a fitted \code{TBATS} model.
}
\details{
This method is used by \code{forecast()} when forecasting a mable containing
a \code{TBATS} model.
}
\examples{
\donttest{
library(dplyr)
library(tsibble)
library(fabletools)

train_frame <- elec_price |>
  filter(bidding_zone == "DE") |>
  slice_head(n = 24 * 21) |>
  as_tsibble(index = time)

model_frame <- train_frame |>
  model("TBATS" = TBATS(value, periods = c(24, 168)))

forecast(model_frame, h = 24)
}
}
\seealso{
Other TBATS: 
\code{\link{TBATS}()},
\code{\link{fitted.TBATS}()},
\code{\link{model_sum.TBATS}()},
\code{\link{residuals.TBATS}()}
}
\concept{TBATS}
