% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/THETA.R
\name{forecast.THETA}
\alias{forecast.THETA}
\title{Forecast a trained THETA model.}
\usage{
\method{forecast}{THETA}(object, new_data, specials = NULL, ...)
}
\arguments{
\item{object}{The trained THETA model used to produce the forecasts.}

\item{new_data}{Forecast horizon.}

\item{specials}{Specials are currently not in use.}

\item{...}{Additional arguments for forecast model method.}
}
\value{
An object of class "fable".
}
\description{
Forecast a trained THETA model.
}
