% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_model.R
\name{forecast_model}
\alias{forecast_model}
\title{Forecast univariate models.}
\usage{
forecast_model(models, n_ahead)
}
\arguments{
\item{models}{A model object of class \code{mdl_df} ("mable"). The result of a call to \code{fabletools::model(...)}.}

\item{n_ahead}{Integer value. The forecast horizon (n-step-ahead).}
}
\value{
A forecast object of class \code{fbl_ts} ("fable") containing the forecasts for all variables, slices and models.
}
\description{
This function takes a model object of class \code{mdl_df} and produces forecasts for a given
   forecast horizon.
}
