% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%out\%}
\alias{\%out\%}
\title{Negated value matching}
\usage{
x \%out\% table
}
\arguments{
\item{x}{Values to be matched.}

\item{table}{Values to be matched against.}
}
\value{
Logical vector, indicating if a non-match was located for each element of x
}
\description{

}
