% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_split.R
\name{initialize_split}
\alias{initialize_split}
\title{Initialize a plan for train-test split}
\usage{
initialize_split(
  main,
  context,
  type = c("first", "last", "prob"),
  value = NULL
)
}
\arguments{
\item{context}{A \code{list} with \code{series_id}, \code{value_id} and \code{index_id}.}

\item{type}{Character value. The type for the initial split.}

\item{value}{Numeric value.}

\item{data}{A \code{tibble} containing the actual values.}
}
\value{
A \code{tibble}
}
\description{
The function creates the initial split into training and
  testing data.
}
