% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{iqr_vec}
\alias{iqr_vec}
\title{Replace outliers based on IQR method by NAs.}
\usage{
iqr_vec(x, alpha = 0.05)
}
\arguments{
\item{x}{Numeric vector.}

\item{alpha}{Numeric value. Controls the width of the limits.}
}
\value{
x Numeric vector.
}
\description{
\code{iqr_vec} takes a numeric vector and replaces
   outliers based on the IQR method by NAs.
}
