% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{kurtosis_vec}
\alias{kurtosis_vec}
\title{Measure of kurtosis.}
\usage{
kurtosis_vec(x, na_rm = TRUE)
}
\arguments{
\item{x}{Numeric vector.}

\item{na_rm}{Logical value. If \code{TRUE}, missing values are dropped.}
}
\value{
Numeric value.
}
\description{
Measure of kurtosis.
}
