% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{log_header}
\alias{log_header}
\title{Create string with header for log file.}
\usage{
log_header()
}
\value{
Character value.
}
\description{
Create a string with information about the current R session
  for the log file.
}
