% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{log_platform}
\alias{log_platform}
\title{Create platform info}
\usage{
log_platform(...)
}
\arguments{
\item{...}{Arguments passed to \code{glue_header()}.}
}
\value{
An object of class \code{glue}.
}
\description{
Create a \code{glue} object with the platform of the
  current R session.
}
