% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lst_to_env}
\alias{lst_to_env}
\title{Assign objects within a list to an environment}
\usage{
lst_to_env(x, envir = .GlobalEnv, ...)
}
\arguments{
\item{x}{A list containing the objects to assign.}

\item{envir}{The environment to use (default is \code{.GlobalEnv}).}

\item{...}{Further arguments passed to \code{assign()}.}
}
\description{
\code{lst_to_env} is a helper function that assigns
  the objects within a list to an environment.
}
