% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{M4_monthly_meta}
\alias{M4_monthly_meta}
\title{Meta data for \code{M4_monthly_data}}
\format{
A \code{tibble} with 30 rows and 8 columns:
   \itemize{
      \item{\code{series_id}: Time series ID from M4 forecasting competition (key variable)}
      \item{\code{category}: Category from M4 forecasting competition (key variable)}
      \item{\code{freq}: Frequency as character}
      \item{\code{n_freq}: Frequency as integer}
      \item{\code{n_ahead}: Forecast horizon as integer}
      \item{\code{n_obs}: Total number of observations}
      \item{\code{start}: Start date}
      \item{\code{end}: End date}
      }
}
\source{
\href{https://github.com/Mcompetitions/M4-methods/tree/master/Dataset}{M4 Competition}
}
\usage{
data(M4_monthly_meta)
}
\description{
The dataset contains meta data for the corresponding
   object \code{M4_monthly_data}.
}
\examples{
data(M4_monthly_meta)
}
\keyword{datasets}
