% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{M4_quarterly_data}
\alias{M4_quarterly_data}
\title{Quarterly time series data from the M4 Competition}
\format{
A time series object of class \code{tsibble} with 7881 rows and 4 columns:
   \itemize{
      \item{\code{date_time}: Date and time (index variable)}
      \item{\code{series_id}: Time series ID from M4 forecasting competition (key variable)}
      \item{\code{category}: Category from M4 forecasting competition (key variable)}
      \item{\code{value}: Time series value (measurement variable)}
      }
}
\source{
\href{https://github.com/Mcompetitions/M4-methods/tree/master/Dataset}{M4 Competition}
}
\usage{
data(M4_quarterly_data)
}
\description{
The dataset contains 30 selected time series on a quarterly basis
   from the M4 Competition.
}
\examples{
data(M4_quarterly_data)
}
\keyword{datasets}
