% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{m5_data}
\alias{m5_data}
\title{Daily Sales Data from the M5-Forecasting Competition}
\format{
A time series object of class \code{tsibble} with 172.170 rows and 10 columns:
   \itemize{
      \item{variable: Unique identifier for category, store and state (key variable)}
      \item{cat_id: Product category (Hobbies, Foods, Household)}
      \item{store_id: Store number}
      \item{state_id: State (CA, TX, WI)}
      \item{date: Date (index variable)}
      \item{year: Year}
      \item{month: Month}
      \item{day: Day}
      \item{wday: Weekday}
      \item{value: Sales data (measurement variable)}
      }
}
\source{
\href{https://www.kaggle.com/c/m5-forecasting-accuracy/data}{M5-Forecasting - Accuracy}
}
\usage{
data(m5_data)
}
\description{
Daily "tsibble" with sales data from the M5-Forecasting
   Competition. The dataset contains time series data
   from 2011-01-29 to 2016-04-24. The sales data are aggregated for
   category (cat_id), stores (store_id) and states (state_id).
}
\examples{
data(m5_data)
}
\keyword{datasets}
