% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{mape_vec}
\alias{mape_vec}
\title{Calculate the mean absolute percentage error}
\usage{
mape_vec(truth, estimate, na_rm = TRUE)
}
\arguments{
\item{truth}{Numeric vector containing the actual values.}

\item{estimate}{Numeric vector containing the forecasts.}

\item{na_rm}{Logical value. If \code{TRUE}, missing values are removed.}
}
\value{
A numeric value.
}
\description{
Calculate the mean absolute percentage error of a numeric vector.
}
\details{
\code{mape_vec()} computes the average absolute percentage forecast error:
\code{abs((truth - estimate) / truth) * 100}.

This metric is undefined when \code{truth} is zero and may return
\code{Inf} or \code{NaN} in such cases.
}
\examples{
truth <- c(10, 20, 40)
estimate <- c(8, 22, 30)

mape_vec(truth, estimate)

truth_na <- c(10, 20, NA)
estimate_na <- c(8, 22, 25)
mape_vec(truth_na, estimate_na)
}
\seealso{
Other accuracy functions: 
\code{\link{mae_vec}()},
\code{\link{make_accuracy}()},
\code{\link{make_errors}()},
\code{\link{me_vec}()},
\code{\link{mpe_vec}()},
\code{\link{mse_vec}()},
\code{\link{rmse_vec}()},
\code{\link{smape_vec}()}
}
\concept{accuracy functions}
