% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{me.data.frame}
\alias{me.data.frame}
\title{Calculate the mean error}
\usage{
\method{me}{data.frame}(data, truth, estimate, na_rm = TRUE, ...)
}
\arguments{
\item{data}{A \code{data.frame}, \code{tibble} or \code{tsibble} containing
the \code{truth} and \code{estimate}.}

\item{truth}{Numeric vector containing the actual values.}

\item{estimate}{Numeric vector containing the forecasts.}

\item{na_rm}{Logical value. If \code{TRUE}, missing values are removed.}

\item{...}{Further arguments passed to \code{metric_summarizer()}.}
}
\value{
A \code{data.frame} or \code{tibble}.
}
\description{
Calculate the mean error for the columns of a data.frame.
   \code{me()} is a metric that is in the same units as the original data.
}
