% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{me}
\alias{me}
\title{Calculate the mean error.}
\usage{
me(data, ...)
}
\arguments{
\item{data}{A `data.frame` containing the `truth` and `estimate`
columns.}

\item{...}{Not currently used.}

\item{truth}{The column identifier for the true results
(that is `numeric`). This should be an unquoted column name although
this argument is passed by expression and supports
[quasiquotation][rlang::quasiquotation] (you can unquote column
names). For `_vec()` functions, a `numeric` vector.}

\item{estimate}{The column identifier for the predicted
results (that is also `numeric`). As with `truth` this can be
specified different ways but the primary method is to use an
unquoted variable name. For `_vec()` functions, a `numeric` vector.}

\item{na_rm}{A  `logical` value indicating whether `NA`
values should be stripped before the computation proceeds.}
}
\value{

}
\description{
Calculate the mean error. `me()` is a metric that is in
the same units as the original data.
}
