% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLP.R
\name{MLP}
\alias{MLP}
\title{Automatic training of MLPs}
\usage{
MLP(formula, ...)
}
\arguments{
\item{formula}{Model specification (see "Specials" section, currently not in use...)}

\item{...}{Further arguments passed to \code{nnfor::mlp()}.}
}
\value{
mlp_model An object of class \code{MLP}.
}
\description{
Automatic training of Multilayer Perceptrons (MLPs). This function
   is a wrapper for \code{nnfor::mlp()}.
}
