% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-fable.R
\name{model_sum.DSHW}
\alias{model_sum.DSHW}
\title{Summarize a DSHW model}
\usage{
\method{model_sum}{DSHW}(x)
}
\arguments{
\item{x}{A fitted \code{DSHW} model object.}
}
\value{
A character string.
}
\description{
Return a short model label for a fitted \code{DSHW} model.
}
\examples{
\donttest{
library(dplyr)
library(tsibble)
library(fabletools)

train_frame <- elec_load |>
  filter(bidding_zone == "DE") |>
  slice_head(n = 24 * 28) |>
  as_tsibble(index = time)

model_frame <- train_frame |>
  model("DSHW" = DSHW(value, periods = c(24, 168)))

model_frame
}
}
\seealso{
Other DSHW: 
\code{\link{DSHW}()},
\code{\link{fitted.DSHW}()},
\code{\link{forecast.DSHW}()},
\code{\link{residuals.DSHW}()}
}
\concept{DSHW}
