% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EXPERT.R
\name{model_sum.EXPERT}
\alias{model_sum.EXPERT}
\title{Provide a succinct summary of a trained EXPERT model}
\usage{
\method{model_sum}{EXPERT}(object)
}
\arguments{
\item{object}{An object of class \code{EXPERT}.}
}
\value{
Model summary as character value.
}
\description{
Provide a succinct summary of a trained EXPERT model.
}
