% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-fable.R
\name{model_sum.MEDIAN}
\alias{model_sum.MEDIAN}
\title{Summarize a median model}
\usage{
\method{model_sum}{MEDIAN}(x)
}
\arguments{
\item{x}{A fitted \code{MEDIAN} model object.}
}
\value{
A character string.
}
\description{
Return a short model label for a fitted \code{MEDIAN} model.
}
\examples{
library(dplyr)
library(tsibble)
library(fabletools)

train_frame <- M4_monthly_data |>
  filter(series == first(series)) |>
  as_tsibble(index = index)

model_frame <- train_frame |>
  model("MEDIAN" = MEDIAN(value ~ window()))

model_frame
}
\seealso{
Other MEDIAN: 
\code{\link{MEDIAN}()},
\code{\link{fitted.MEDIAN}()},
\code{\link{forecast.MEDIAN}()},
\code{\link{residuals.MEDIAN}()}
}
\concept{MEDIAN}
