% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-fable.R
\name{model_sum.NAIVE2}
\alias{model_sum.NAIVE2}
\title{Summarize a Naive2 model}
\usage{
\method{model_sum}{NAIVE2}(x)
}
\arguments{
\item{x}{A fitted \code{NAIVE2} model object.}
}
\value{
A character string.
}
\description{
Return a short model label for a fitted \code{NAIVE2} model.
}
\examples{
library(dplyr)
library(tsibble)
library(fabletools)

train_frame <- M4_monthly_data |>
  filter(series == first(series)) |>
  as_tsibble(index = index)

model_frame <- train_frame |>
  model("NAIVE2" = NAIVE2(value ~ season(12)))

model_frame
}
\seealso{
Other NAIVE2: 
\code{\link{NAIVE2}()},
\code{\link{fitted.NAIVE2}()},
\code{\link{forecast.NAIVE2}()},
\code{\link{residuals.NAIVE2}()}
}
\concept{NAIVE2}
