% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-fable.R
\name{model_sum.SMEAN}
\alias{model_sum.SMEAN}
\title{Summarize a seasonal mean model}
\usage{
\method{model_sum}{SMEAN}(x)
}
\arguments{
\item{x}{A fitted \code{SMEAN} model object.}
}
\value{
A character string.
}
\description{
Return a short model label for a fitted \code{SMEAN} model.
}
\examples{
library(dplyr)
library(tsibble)
library(fabletools)

train_frame <- M4_monthly_data |>
  filter(series == first(series)) |>
  as_tsibble(index = index)

model_frame <- train_frame |>
  model("SMEAN" = SMEAN(value ~ lag("year")))

model_frame
}
\seealso{
Other SMEAN: 
\code{\link{SMEAN}()},
\code{\link{fitted.SMEAN}()},
\code{\link{forecast.SMEAN}()},
\code{\link{residuals.SMEAN}()}
}
\concept{SMEAN}
