% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/THETA.R
\name{model_sum.THETA}
\alias{model_sum.THETA}
\title{Provide a succinct summary of the THETA.}
\usage{
\method{model_sum}{THETA}(x)
}
\arguments{
\item{object}{The THETA to summarize.}
}
\value{

}
\description{
Provide a succinct summary of the THETA.
}
