% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{mpe}
\alias{mpe}
\title{Calculate the mean percentage error}
\usage{
mpe(data, ...)
}
\arguments{
\item{data}{A \code{data.frame}, \code{tibble} or \code{tsibble} containing
the \code{truth} and \code{estimate}.}

\item{...}{Not currently used.}
}
\value{

}
\description{
Calculate the mean percentage error. \code{mpe()} is a metric
   that is in relative units.
}
