% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{name_file}
\alias{name_file}
\title{Create a name for a folder or file}
\usage{
name_file(
  primary,
  extension = NULL,
  add_time = TRUE,
  time_format = "\%Y\%m\%d_\%H\%M\%S"
)
}
\arguments{
\item{primary}{Character value. Primary file name to be used.}

\item{extension}{Character value. File name extension to be used
(e.g. \code{".csv"}, \code{".rda"}).}

\item{add_time}{Logical value. If \code{TRUE}, a time stamp is added in
front of the file name.}

\item{time_format}{Character value. The time format of the time stamp.}
}
\value{
Character value.
}
\description{
The function creates a name for a folder or file.
}
