% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{number_string}
\alias{number_string}
\title{Helper function to create numbered strings.}
\usage{
number_string(x, n)
}
\arguments{
\item{x}{Character value.}

\item{n}{Integer value.}
}
\value{
x Character vector.
}
\description{
Creates a character vector in the form
  \code{c("x(1)", x(2), ..., x(n))}.
}
