% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_data.R
\name{pacf_vec}
\alias{pacf_vec}
\title{Estimate partial autocorrelations of a numeric vector}
\usage{
pacf_vec(x, lag_max = 24, ...)
}
\arguments{
\item{x}{Numeric vector.}

\item{lag_max}{Integer. Maximum lag for which the partial autocorrelation is
estimated.}

\item{...}{Further arguments passed to \code{stats::pacf()}.}
}
\value{
A numeric vector containing the sample partial autocorrelations for lags
\code{1} to \code{lag_max}.
}
\description{
Estimate the sample partial autocorrelation function of a numeric vector.
}
\details{
\code{pacf_vec()} is a small wrapper around \code{stats::pacf()}. It returns
the sample partial autocorrelations as a numeric vector for lags
\code{1} to \code{lag_max}.
}
\examples{
library(dplyr)

x <- M4_monthly_data |>
  filter(series == first(series)) |>
  pull(value)

pacf_vec(
  x = x,
  lag_max = 12
)
}
\seealso{
Other data analysis: 
\code{\link{acf_vec}()},
\code{\link{estimate_acf}()},
\code{\link{estimate_kurtosis}()},
\code{\link{estimate_mode}()},
\code{\link{estimate_pacf}()},
\code{\link{estimate_skewness}()},
\code{\link{summarise_data}()},
\code{\link{summarise_split}()},
\code{\link{summarise_stats}()}
}
\concept{data analysis}
