% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pct_errors.R
\name{pct_errors}
\alias{pct_errors}
\title{Calculate percentage forecast errors}
\usage{
pct_errors(fcst, test)
}
\arguments{
\item{fcst}{A \code{fable} containing the forecasts for the models, splits, etc.}

\item{test}{A \code{tsibble} containing the testing data.}
}
\value{
errors A \code{tsibble}.
}
\description{
\code{pct_error} calculates percentage forecast errors (actual - fcst / actual) * 100.
}
