% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_acf.R
\name{plot_acf}
\alias{plot_acf}
\title{Plot autocorrelation function.}
\usage{
plot_acf(
  data,
  lag_max = 25,
  demean = TRUE,
  level = 0.9,
  title = "Correlation analysis",
  subtitle = "Sample autocorrelation (ACF)",
  xlab = "Lag k",
  ylab = NULL,
  caption = NULL,
  bar_width = 1,
  bar_color = "#31688EFF",
  bar_color2 = "#D55E00",
  bar_alpha = 0.6,
  line_width = 0.25,
  line_type = "solid",
  line_color = "grey35",
  theme_set = theme_gray(),
  theme_config = list()
)
}
\arguments{
\item{data}{A valid tsibble in long format with one measurement variable.}

\item{lag_max}{Integer value. Maximum number of lags.}

\item{demean}{Logical value. If \code{TRUE}, the time series is demeaned.}

\item{level}{Numeric value. The confidence level to check significance.}

\item{title}{Title of the plot.}

\item{subtitle}{Subtitle of the plot.}

\item{xlab}{Label for the x-axis.}

\item{ylab}{Label for the y-axis.}

\item{caption}{Caption of the plot.}

\item{bar_width}{Numeric value. The width of the bars.}

\item{bar_color}{Character value. The color of the significant bars.}

\item{bar_color2}{Character value. The color of the non-significant bars.}

\item{bar_alpha}{Numeric value. The transparency of the bars.}

\item{line_width}{Numeric value. Line width of the confidence line.}

\item{line_type}{Character value. Line type of the confidence line.}

\item{line_color}{Character value. Line color of the confidence line.}

\item{theme_set}{A complete ggplot2 theme.}

\item{theme_config}{A list with further arguments passed to \code{ggplot2::theme()}.}
}
\value{
p An object of class ggplot.
}
\description{
This function plots the sample autocorrelation function of one or more time series.
}
