% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dist.R
\name{plot_dist}
\alias{plot_dist}
\title{plot_dist}
\usage{
plot_dist(
  data,
  legend = FALSE,
  legend_position = "right",
  title = NULL,
  subtitle = NULL,
  xlab = NULL,
  ylab = "Density",
  caption = NULL,
  bar_color = "#31688EFF",
  bar_alpha = 0.3,
  line_width = 0.1,
  line_type = "solid",
  line_color = "#31688EFF",
  line_fill = "#31688EFF"
)
}
\arguments{
\item{data}{A tsibble containing the columns time, variable and value.}

\item{legend}{Logical value. If TRUE, a legend is displayed.}

\item{legend_position}{Character value defining the position of the legend ("top", "bottom", "right", "left").}

\item{title}{Title of the plot.}

\item{subtitle}{Subtitle of the plot.}

\item{xlab}{Label for the x-axis.}

\item{ylab}{Label for the y-axis.}

\item{caption}{Caption of the plot.}

\item{bar_color}{Character value defining the color of the histrogram bars}

\item{bar_alpha}{Numeric value defining the transparency of the histogram bars}

\item{line_width}{Numeric value defining the line width of the kernel density estimator}

\item{line_type}{Integer value defining the line type of the kernel density estimator}

\item{line_color}{Character value defining the line color of the kernel density estimator}
}
\value{
p An object of class ggplot
}
\description{
Plot the distribution of a time series (histogram, kernel density estimate, normal distribution).
}
