% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_error_dist.R
\name{plot_error_dist}
\alias{plot_error_dist}
\title{Plot the forecast errors along the forecast horizon.}
\usage{
plot_error_dist(
  data,
  variable,
  model,
  add_points = TRUE,
  jitter = TRUE,
  title = variable,
  subtitle = "Distribution of forecast errors by method",
  xlab = "Forecast horizon (n-step)",
  ylab = "Forecast error",
  caption = NULL,
  point_size = 2,
  point_shape = 19,
  point_alpha = 0.25,
  box_line_width = 0.25,
  box_alpha = 0.75,
  base_size = 11
)
}
\arguments{
\item{data}{A tsibble containing the forecasts errors for all target variables, slice, models etc. provided by the function \code{create_errors(...)}.}

\item{variable}{Character value. The target variable.}

\item{model}{Character vector. Choose the forecasting methods (e.g. c("ARIMA", "ETS"))}

\item{add_points}{Logical value. If \code{TRUE}, the errors are plotted as points. Ohterwise, Box-Whisker plots only.}

\item{jitter}{Logical value. If \code{TRUE}, points are jittered.}

\item{title}{Title for the plot}

\item{subtitle}{Subtitle for the plot}

\item{xlab}{Label for the x-axis}

\item{ylab}{Label for the y-axis}

\item{caption}{Caption of the plot.}

\item{point_size}{Numeric value defining the point size.}

\item{point_alpha}{Numeric value. The transparency of the points.}

\item{box_line_width}{Numeric value defining the line width of the Box-Whisker.}

\item{box_alpha}{Numeric value. The transparency of the Box-Whisker.}

\item{base_size}{Integer value. Base font size.}

\item{pointshape}{Integer value defining the point shape.}
}
\value{
p An object of class ggplot
}
\description{
Plot the forecast errors along the forecast horizon (and corresponding Box-Whisker Plots) for a specific target variable and forecasting methods.
}
