% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_error_metrics.R
\name{plot_error_metrics}
\alias{plot_error_metrics}
\title{Plot forecast accuracy metrics.}
\usage{
plot_error_metrics(
  data,
  variable,
  model,
  metric,
  dim = "horizon",
  title = "Evaluation of forecast accuracy",
  ylab = NULL,
  caption = NULL,
  line_width = 1.5,
  point_size = 3,
  point_shape = 19,
  point_alpha = 1,
  base_size = 11
)
}
\arguments{
\item{data}{A tibble containing the accuracy metrics, i.e. the result of a call to \code{error_metrics(...)}.}

\item{variable}{Character vector defining the target variable.}

\item{model}{Character vector defining the forecasting methods.}

\item{metric}{Character vector defining the accuracy measures.}

\item{title}{Title for the plot.}

\item{ylab}{Label for the y-axis.}

\item{caption}{Caption (footnote) for the plot.}

\item{line_width}{Numeric value. Line width.}

\item{point_size}{Numeric value. Point size.}

\item{point_shape}{Integer value. Point shape.}

\item{point_alpha}{Numeric value between 0 and 1. The transparency of the points.}

\item{base_size}{Integer value. Base font size.}

\item{subtitle}{Subtitle for the plot.}

\item{xlab}{Label for the x-axis.}
}
\value{
p An object of class ggplot.
}
\description{
Plot forecast accuracy metrics, either along the forecast horizon or the slices. The user can define
   specific target variable(s), forecasting method(s) and accuracy metric(s).
}
