% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_forecast.R
\name{plot_forecast}
\alias{plot_forecast}
\title{Plot the actual and fitted values and forecasts.}
\usage{
plot_forecast(
  data,
  variable = NULL,
  slice = NULL,
  model = NULL,
  include = 24,
  title = NULL,
  subtitle = NULL,
  xlab = "Date",
  ylab = NULL,
  caption = NULL,
  line_width = 1.5,
  point_size = 2,
  point_alpha = 0,
  base_size = 11
)
}
\arguments{
\item{data}{A tsibble according to the adice data model.}

\item{variable}{Character vector defining the variable(s).}

\item{slice}{Integer vector. The time slice of the rolling forecast.}

\item{model}{Character vector. The forecasting methods.}

\item{include}{Integer value. The number of actual values (training) to be included.}

\item{title}{Title of the plot.}

\item{subtitle}{Subtitle of the plot.}

\item{xlab}{Label for the x-axis.}

\item{ylab}{Label for the y-axis.}

\item{caption}{Caption of the plot.}

\item{line_width}{Numeric value defining the line width.}

\item{point_size}{Numeric value defining the point size.}

\item{point_alpha}{Numeric value. The transparency of the points.}

\item{base_size}{Integer value. Base font size.}
}
\value{
p An object of class ggplot.
}
\description{
Plot the actual and fitted values and forecasts for specific target variable(s), slice(s) and forecasting method(s).
}
