% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pacf.R
\name{plot_pacf}
\alias{plot_pacf}
\title{Plot the sample partial autocorrelation function (PACF).}
\usage{
plot_pacf(
  data,
  lag_max = 24,
  level = 0.9,
  title = NULL,
  subtitle = NULL,
  xlab = "Lag k",
  ylab = "PACF",
  caption = NULL,
  bar_width = 1,
  bar_color = "#31688EFF",
  bar_alpha = 0.6,
  line_width = 0.25,
  line_type = "solid",
  line_color = "grey35",
  legend = FALSE,
  legend_position = "right"
)
}
\arguments{
\item{data}{A tsibble containing the columns time, variable and value.}

\item{lag_max}{Integer value defining the maximum number of lags.}

\item{level}{Numeric value defining the confidence level.}

\item{title}{Title of the plot.}

\item{subtitle}{Subtitle of the plot.}

\item{xlab}{Label for the x-axis.}

\item{ylab}{Label for the y-axis.}

\item{caption}{Caption of the plot.}

\item{bar_width}{Numeric value defining the bar width.}

\item{bar_color}{Character value defining the bar color.}

\item{line_width}{Numeric value defining the line width.}

\item{line_type}{Integer value defining the line type.}

\item{line_color}{Character value defining the line color.}

\item{legend}{Logical value. If \code{TRUE}, a legend is plotted.}

\item{legend_position}{Character value defining the position of the legend ("top", "bottom", "right", "left").}
}
\value{
p An object of class ggplot.
}
\description{
Bar chart of the sample partial autocorrelation function (PACF) of a time series.
}
