% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_season.R
\name{plot_season}
\alias{plot_season}
\title{Seasonal plot of a time series.}
\usage{
plot_season(
  data,
  title = "Seasonal plot",
  subtitle = "Median and quantiles",
  xlab = "Hour",
  ylab = NULL,
  caption = NULL,
  line_width = 0.75,
  line_type = "solid",
  line_color = "#31688EFF",
  line_alpha = 1,
  base_size = 10
)
}
\arguments{
\item{data}{A tsibble containing the columns time, variable and value.}

\item{title}{Title of the plot.}

\item{subtitle}{Subtitle of the plot.}

\item{xlab}{Label for the x-axis.}

\item{ylab}{Label for the y-axis.}

\item{caption}{Caption of the plot.}

\item{line_width}{Numeric value defining the line width (time series).}

\item{line_type}{Numeric value defining the line type (time series).}

\item{line_color}{Character value defining the line color (time series).}

\item{line_alpha}{Numeric value defining the transparency of the line.}

\item{base_size}{Integer value. Base font size.}
}
\value{
p An object of class ggplot.
}
\description{
Seasonal plot of one or more time series.
}
