% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ts.R
\name{plot_ts}
\alias{plot_ts}
\title{Plot time series as line chart.}
\usage{
plot_ts(
  data,
  title = NULL,
  subtitle = NULL,
  xlab = NULL,
  ylab = NULL,
  caption = NULL,
  smooth_width = 0.5,
  smooth_type = "solid",
  smooth_color = "#D55E00",
  line_width = 0.75,
  line_type = "solid",
  line_color = "#31688EFF",
  line_alpha = 1,
  smooth_method = NULL,
  smooth_fill = "#D55E00",
  smooth_alpha = 0.25,
  theme_set = theme_tscv(),
  theme_config = list(),
  ...
)
}
\arguments{
\item{data}{A valid tsibble in long format with one measurement variable.}

\item{title}{Title of the plot.}

\item{subtitle}{Subtitle of the plot.}

\item{xlab}{Label for the x-axis.}

\item{ylab}{Label for the y-axis.}

\item{caption}{Caption of the plot.}

\item{smooth_width}{Numeric value defining the line width (smooth curve).}

\item{smooth_type}{Numeric value defining the line type (smooth curve).}

\item{smooth_color}{Character value defining the line color (smooth curve).}

\item{line_width}{Numeric value defining the line width (time series).}

\item{line_type}{Numeric value defining the line type (time series).}

\item{line_color}{Character value defining the line color (time series).}

\item{smooth_method}{Character value. The smoothing method (\code{NULL}, \code{"loess"}, \code{"lm"}, \code{glm}, \code{gam}).}

\item{smooth_fill}{Character value defining the fill color of confidence band.}

\item{smooth_alpha}{Numeric value defining the transparency of confidence band.}

\item{theme_set}{A complete ggplot2 theme.}

\item{theme_config}{A list with further arguments passed to \code{ggplot2::theme()}.}

\item{...}{Further arguments passed to \code{ggplot2::stat_smooth()}}

\item{smooth_se}{Logical value. If \code{TRUE}, confidence interval is displayed around smooth.}
}
\value{
p An object of class ggplot.
}
\description{
Plot one or more time series incl. smooth curve (linear or loess) as line chart.
}
