% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_split.R
\name{print_split}
\alias{print_split}
\title{Print the split table as \code{gt} object.}
\usage{
print_split(data, include = 3)
}
\arguments{
\item{data}{A tibble given by the function \code{summarise_split(...)}.}

\item{include}{Integer value. The number of top and bottom rows to include. If \code{NULL}, all rows are included.}
}
\value{
An object of class \code{gt}.
}
\description{
This function prints the split table as \code{gt} object.
}
