% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-fable.R
\name{residuals.DSHW}
\alias{residuals.DSHW}
\title{Extract residuals from a DSHW model}
\usage{
\method{residuals}{DSHW}(object, ...)
}
\arguments{
\item{object}{A fitted \code{DSHW} model object.}

\item{...}{Additional arguments. Currently not used.}
}
\value{
Residuals.
}
\description{
Extract residuals from a fitted \code{DSHW} model.
}
\examples{
\donttest{
library(dplyr)
library(tsibble)
library(fabletools)

train_frame <- elec_price |>
  filter(bidding_zone == "DE") |>
  slice_head(n = 24 * 28) |>
  as_tsibble(index = time)

model_frame <- train_frame |>
  model("DSHW" = DSHW(value, periods = c(24, 168)))

residuals(model_frame)
}
}
\seealso{
Other DSHW: 
\code{\link{DSHW}()},
\code{\link{fitted.DSHW}()},
\code{\link{forecast.DSHW}()},
\code{\link{model_sum.DSHW}()}
}
\concept{DSHW}
