% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELM.R
\name{residuals.ELM}
\alias{residuals.ELM}
\title{Extract residuals from a trained ELM model}
\usage{
\method{residuals}{ELM}(object, ...)
}
\arguments{
\item{object}{An object of class \code{ELM}.}

\item{...}{Currently not in use.}
}
\value{
Fitted values as tsibble.
}
\description{
Extract residuals from a trained ELM model.
}
