% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EXPERT.R
\name{residuals.EXPERT}
\alias{residuals.EXPERT}
\title{Extract residuals from a trained EXPERT model}
\usage{
\method{residuals}{EXPERT}(object, ...)
}
\arguments{
\item{object}{An object of class \code{EXPERT}.}

\item{...}{Currently not in use.}
}
\value{
Fitted values as tsibble.
}
\description{
Extract residuals from a trained EXPERT model.
}
