% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-fable.R
\name{residuals.MEDIAN}
\alias{residuals.MEDIAN}
\title{Extract residuals from a median model}
\usage{
\method{residuals}{MEDIAN}(object, ...)
}
\arguments{
\item{object}{A fitted \code{MEDIAN} model object.}

\item{...}{Additional arguments. Currently not used.}
}
\value{
Residuals.
}
\description{
Extract residuals from a fitted \code{MEDIAN} model.
}
\examples{
library(dplyr)
library(tsibble)
library(fabletools)

train_frame <- M4_monthly_data |>
  filter(series == first(series)) |>
  as_tsibble(index = index)

model_frame <- train_frame |>
  model("MEDIAN" = MEDIAN(value ~ window()))

residuals(model_frame)
}
\seealso{
Other MEDIAN: 
\code{\link{MEDIAN}()},
\code{\link{fitted.MEDIAN}()},
\code{\link{forecast.MEDIAN}()},
\code{\link{model_sum.MEDIAN}()}
}
\concept{MEDIAN}
