% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLP.R
\name{residuals.MLP}
\alias{residuals.MLP}
\title{Extract residuals from a trained MLP model}
\usage{
\method{residuals}{MLP}(object, ...)
}
\arguments{
\item{object}{An object of class \code{MLP}.}

\item{...}{Currently not in use.}
}
\value{
Fitted values as tsibble.
}
\description{
Extract residuals from a trained MLP model.
}
