% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-fable.R
\name{residuals.NAIVE2}
\alias{residuals.NAIVE2}
\title{Extract residuals from a Naive2 model}
\usage{
\method{residuals}{NAIVE2}(object, ...)
}
\arguments{
\item{object}{A fitted \code{NAIVE2} model object.}

\item{...}{Additional arguments. Currently not used.}
}
\value{
Residuals.
}
\description{
Extract residuals from a fitted \code{NAIVE2} model.
}
\examples{
library(dplyr)
library(tsibble)
library(fabletools)

train_frame <- M4_monthly_data |>
  filter(series == first(series)) |>
  as_tsibble(index = index)

model_frame <- train_frame |>
  model("NAIVE2" = NAIVE2(value ~ season(12)))

residuals(model_frame)
}
\seealso{
Other NAIVE2: 
\code{\link{NAIVE2}()},
\code{\link{fitted.NAIVE2}()},
\code{\link{forecast.NAIVE2}()},
\code{\link{model_sum.NAIVE2}()}
}
\concept{NAIVE2}
