% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-fable.R
\name{residuals.SMEAN}
\alias{residuals.SMEAN}
\title{Extract residuals from a seasonal mean model}
\usage{
\method{residuals}{SMEAN}(object, ...)
}
\arguments{
\item{object}{A fitted \code{SMEAN} model object.}

\item{...}{Additional arguments. Currently not used.}
}
\value{
Residuals.
}
\description{
Extract residuals from a fitted \code{SMEAN} model.
}
\examples{
library(dplyr)
library(tsibble)
library(fabletools)

train_frame <- M4_monthly_data |>
  filter(series == first(series)) |>
  as_tsibble(index = index)

model_frame <- train_frame |>
  model("SMEAN" = SMEAN(value ~ lag("year")))

residuals(model_frame)
}
\seealso{
Other SMEAN: 
\code{\link{SMEAN}()},
\code{\link{fitted.SMEAN}()},
\code{\link{forecast.SMEAN}()},
\code{\link{model_sum.SMEAN}()}
}
\concept{SMEAN}
