% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-fable.R
\name{residuals.SNAIVE2}
\alias{residuals.SNAIVE2}
\title{Extract residuals from a SNAIVE2 model}
\usage{
\method{residuals}{SNAIVE2}(object, ...)
}
\arguments{
\item{object}{A fitted \code{SNAIVE2} model object.}

\item{...}{Additional arguments. Currently not used.}
}
\value{
Residuals.
}
\description{
Extract residuals from a fitted \code{SNAIVE2} model.
}
\examples{
library(dplyr)
library(tsibble)
library(fabletools)

train_frame <- elec_price |>
  filter(bidding_zone == "DE") |>
  slice_head(n = 24 * 21) |>
  as_tsibble(index = time)

model_frame <- train_frame |>
  model("SNAIVE2" = SNAIVE2(value))

residuals(model_frame)
}
\seealso{
Other SNAIVE2: 
\code{\link{SNAIVE2}()},
\code{\link{fitted.SNAIVE2}()},
\code{\link{forecast.SNAIVE2}()},
\code{\link{model_sum.SNAIVE2}()}
}
\concept{SNAIVE2}
