% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/THETA.R
\name{residuals.THETA}
\alias{residuals.THETA}
\title{Extract residuals from a THETA.}
\usage{
\method{residuals}{THETA}(object, ...)
}
\arguments{
\item{object}{The time series model used to produce the forecast.}

\item{...}{Further arguments.}
}
\value{

}
\description{
Extract residuals from a THETA.
}
