% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{response_vars}
\alias{response_vars}
\title{Return response variables.}
\usage{
response_vars(.data)
}
\arguments{
\item{.data}{A \code{tsibble} or \code{fable}.}
}
\value{
response A character vector.
}
\description{
\code{response_vars()} returns a character vector with the response variables, i.e.
   key variables without "helper variables" like \code{split}, \code{id}, \code{sample}, etc.
}
