% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{rmse_vec}
\alias{rmse_vec}
\title{Calculate the root mean squared error}
\usage{
rmse_vec(truth, estimate, na_rm = TRUE)
}
\arguments{
\item{truth}{Numeric vector containing the actual values.}

\item{estimate}{Numeric vector containing the forecasts.}

\item{na_rm}{Logical value. If \code{TRUE}, missing values are removed.}
}
\value{
A numeric value.
}
\description{
Calculate the root mean squared error of a numeric vector.
}
\details{
\code{rmse_vec()} computes the square root of the mean squared forecast error.
The metric is reported in the same units as the original data.
}
\examples{
truth <- c(10, 20, 30)
estimate <- c(8, 22, 25)

rmse_vec(truth, estimate)

truth_na <- c(10, 20, NA)
estimate_na <- c(8, 22, 25)
rmse_vec(truth_na, estimate_na)
}
\seealso{
Other accuracy functions: 
\code{\link{mae_vec}()},
\code{\link{make_accuracy}()},
\code{\link{make_errors}()},
\code{\link{mape_vec}()},
\code{\link{me_vec}()},
\code{\link{mpe_vec}()},
\code{\link{mse_vec}()},
\code{\link{smape_vec}()}
}
\concept{accuracy functions}
