% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors_tscv.R
\name{scale_fill_tscv}
\alias{scale_fill_tscv}
\title{Create a tscv fill scale}
\usage{
scale_fill_tscv(palette = "main", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character value. Name of the palette.}

\item{discrete}{Logical value. If \code{TRUE}, create a discrete fill scale.
If \code{FALSE}, create a continuous fill scale.}

\item{reverse}{Logical value. If \code{TRUE}, the palette is reversed.}

\item{...}{Additional arguments passed to \code{ggplot2::discrete_scale()} or
\code{ggplot2::scale_fill_gradientn()}.}
}
\value{
A \code{ggplot2} scale object.
}
\description{
Create a \code{ggplot2} fill scale based on a predefined \code{tscv} palette.
}
\details{
\code{scale_fill_tscv()} creates either a discrete or continuous fill scale
for the \code{fill} aesthetic.

For discrete variables, the function uses \code{ggplot2::discrete_scale()}.
For continuous variables, it uses \code{ggplot2::scale_fill_gradientn()}.

Available palettes are \code{"main"}, \code{"cool"}, \code{"hot"},
\code{"mixed"}, and \code{"grey"}.
}
\examples{
library(dplyr)

context <- list(
  series_id = "series",
  value_id = "value",
  index_id = "index"
)

data <- M4_monthly_data |>
  filter(series \%in\% c("M23100", "M14395"))

stats <- summarise_stats(
  .data = data,
  context = context
)

plot_bar(
  data = stats,
  x = series,
  y = mean,
  color = series,
  title = "Average Value by Series",
  xlab = "Series",
  ylab = "Mean"
) +
  scale_fill_tscv(palette = "main")
}
\seealso{
Other data visualization: 
\code{\link{plot_bar}()},
\code{\link{plot_density}()},
\code{\link{plot_histogram}()},
\code{\link{plot_line}()},
\code{\link{plot_point}()},
\code{\link{plot_qq}()},
\code{\link{scale_color_tscv}()},
\code{\link{theme_tscv}()},
\code{\link{tscv_cols}()},
\code{\link{tscv_pal}()}
}
\concept{data visualization}
