% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{skewness_vec}
\alias{skewness_vec}
\title{Measure of skewness.}
\usage{
skewness_vec(x, na_rm = TRUE)
}
\arguments{
\item{x}{Numeric vector.}

\item{na_rm}{Logical value. If \code{TRUE}, missing values are dropped.}
}
\value{
Numeric value.
}
\description{
Measure of skewness.
}
