% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-fable.R
\name{SMEDIAN}
\alias{SMEDIAN}
\title{Seasonal median model}
\usage{
SMEDIAN(formula, ...)
}
\arguments{
\item{formula}{A model formula specifying the response and \code{lag()}
special, for example \code{value ~ lag("week")}.}

\item{...}{Further arguments.}
}
\value{
A model definition that can be used inside \code{fabletools::model()}.
}
\description{
Specify a seasonal median benchmark model for use with
\code{fabletools::model()}.
}
\details{
\code{SMEDIAN()} forecasts each future observation using the historical median
of the matching seasonal position. Use the \code{lag()} special to define the
seasonal period, for example \code{lag("week")} for hourly data with weekly
seasonality.
}
\examples{
library(dplyr)
library(tsibble)
library(fabletools)

train_frame <- elec_price |>
  filter(bidding_zone == "DE") |>
  slice_head(n = 24 * 21) |>
  as_tsibble(index = time)

model_frame <- train_frame |>
  model("SMEDIAN" = SMEDIAN(value ~ lag("week")))

model_frame
}
\seealso{
Other SMEDIAN: 
\code{\link{fitted.SMEDIAN}()},
\code{\link{forecast.SMEDIAN}()},
\code{\link{model_sum.SMEDIAN}()},
\code{\link{residuals.SMEDIAN}()}
}
\concept{SMEDIAN}
