% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_data.R
\name{split_data}
\alias{split_data}
\title{Split the data into training and testing.}
\usage{
split_data(data, mode = "stretch", n_init, n_step, n_ahead)
}
\arguments{
\item{data}{A tsibble in long format with standard columns (date_time, variable and value) provided by the function \code{clean_data(...)}.}

\item{mode}{Character value. Define the setup of the training window for time series cross validation. \code{stretch} is equivalent to an expanding window approach and \code{slide} is a fixed window approach.}

\item{n_init}{Integer value. The number of periods for the initial training window (must be positive).}

\item{n_step}{Integer value. The number of periods to skip between windows (must be positive).}

\item{n_ahead}{Integer value. The forecast horizon (n-steps-ahead, must be positive).}
}
\value{
data A tsibble with the same format like the input data, but with additional columns:
   \itemize{
      \item{\code{slice}: Integer value. The number of the time slice (training and testing).}
      \item{\code{sample}: Character value. Indicating whether the partition is training or testing.}
      \item{\code{horizon}: Integer value. The forecast horizon (i.e. the size of the testing window. NA for training.}
      \item{\code{id}: Integer value. The number of the observations within training and testing.}
      }
}
\description{
This function splits the data into training and testing (i.e. partitioning into time slices) for
   time series cross-validation. The user can choose between \code{stretch} and \code{slide}. The first is an
   expanding window approach, while the latter is a fixed window approach. The user can define the window sizes
   for training and testing via \code{n_init} and \code{n_ahead}, as well as the step size for increments via
   \code{n_step}.
}
