% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack_data.R
\name{stack_data}
\alias{stack_data}
\title{Stack a tsibble and a fable.}
\usage{
stack_data(data, fcsts, models)
}
\arguments{
\item{data}{A tsibble in long format with standard columns (.date, .variable, .value, etc.) provided by the function \code{slice_data(...)}.}

\item{fcsts}{A forecast object of class \code{fbl_ts} (i.e. "fable" or "forecast table").}

\item{models}{A model object of class \code{mdl_ts} (i.e. "mable" or "model table").}
}
\value{
data A tsibble in long format with standard columns (.date, .variable, .value, etc.). Further columns are added ...
}
\description{
This function "stacks" the historical from the object data (tsibble), the forecasts from object fcsts
   (fable) and the fitted values from object models (mable). Note: The column \code{.distribution} from the forecast
   object ist dropped. This is primary a helper function ...
}
