% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{target_vars}
\alias{target_vars}
\title{Return target variables}
\usage{
target_vars(.data)
}
\arguments{
\item{.data}{A \code{tsibble} or \code{fable}.}
}
\value{
target A character vector.
}
\description{
\code{target_vars()} returns a character vector with the target variables, i.e.
   key variables without "helper variables" like \code{.split}, \code{.id}, \code{.sample}, etc.
}
