% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_tscv_dark.R
\name{theme_tscv_dark}
\alias{theme_tscv_dark}
\title{Dark ggplot2 theme for tscv package}
\usage{
theme_tscv_dark(
  base_size = 11,
  base_family = "",
  base_line_size = base_size/22,
  base_rect_size = base_size/22
)
}
\arguments{
\item{base_size}{Base font size.}

\item{base_family}{Base font family.}

\item{base_line_size}{Base size for "line" elements.}

\item{base_rect_size}{Base size for "rect" elements.}
}
\description{
Custom ggplot2 theme for tscv package.
}
