% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/THETA.R
\name{THETA}
\alias{THETA}
\title{Automatic training of THETA.}
\usage{
THETA(formula, ...)
}
\arguments{
\item{formula}{Model specification (see "Specials" section, currently not in use...)}

\item{...}{Further arguments passed to \code{forecast::theta()}.}
}
\value{
theta_model An object of class \code{THETA}.
}
\description{
Automatic training a THETA model. This function
   is a wrapper for \code{forecast::thetaf()}.
}
