% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSHW.R
\name{train_dshw}
\alias{train_dshw}
\title{Train DSHW model.}
\usage{
train_dshw(.data, specials, ...)
}
\arguments{
\item{.data}{Input data as \code{tsibble}.}

\item{specials}{Specials as list defined in \code{specials_dshw}.}

\item{...}{Further arguments passed to \code{forecast::dshw()}.}
}
\value{
An object of class \code{DSHW}.
}
\description{
Train a Double Seasonal Holt-Winters model (DSHW).
}
