% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELM.R
\name{train_elm}
\alias{train_elm}
\title{Train ELM neural network}
\usage{
train_elm(.data, specials, ...)
}
\arguments{
\item{.data}{Input data as \code{tsibble}.}

\item{specials}{Specials as list defined in \code{specials_elm}.}

\item{...}{Further arguments passed to \code{nnfor::elm()}.}
}
\value{
An object of class \code{ELM}.
}
\description{
Train an Extreme Learning Machine (ELM).
}
