% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EXPERT.R
\name{train_expert}
\alias{train_expert}
\title{EXPERT model}
\usage{
train_expert(.data, specials = NULL, periods, xreg = NULL, ...)
}
\arguments{
\item{.data}{Input data as tsibble.}

\item{specials}{Specials as list defined in \code{specials_expert}.}

\item{periods}{Integer vector. The periodicity of the time series (e.g. \code{periods = c(24, 168)} for hourly data).}

\item{xreg}{A \code{tsibble} containing exogenous variables.}

\item{...}{Further arguments (currently not in use).}
}
\value{
An object of class \code{EXPERT}.
}
\description{
Train the EXPERT model.
}
