% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEDIAN.R
\name{train_median}
\alias{train_median}
\title{Median model}
\usage{
train_median(.data, specials, ...)
}
\arguments{
\item{.data}{Input data as tsibble.}

\item{specials}{Specials as list defined in \code{specials_median}.}

\item{...}{Currently not in use.}
}
\value{
An object of class \code{MEDIAN}.
}
\description{
Train median model (MEDIAN).
}
