% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLP.R
\name{train_mlp}
\alias{train_mlp}
\title{Multilayer Perceptron (MLP)}
\usage{
train_mlp(.data, specials, ...)
}
\arguments{
\item{.data}{Input data as tsibble.}

\item{specials}{Specials as list defined in \code{specials_mlp}.}

\item{...}{Further arguments passed to \code{nnfor::mlp()}.}
}
\value{
An object of class \code{MLP}.
}
\description{
Train a Multilayer Perceptron (MLP) model.
}
