% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_model.R
\name{train_model}
\alias{train_model}
\title{Train univariate models.}
\usage{
train_model(data)
}
\arguments{
\item{data}{A tsibble in long format with standard columns (.date, .variable, .value, etc.).}
}
\value{
models A model object of class \code{mdl_df} ("mable").
}
\description{
This function trains several univariate models to the key variables within data (.variable and .slice).
   The user can define the models via \code{methods}. Currently, the following methods are available:
   \itemize{
      \item{sNaive: Set each forecast to be equal to the last observed value from the same season of the previous period.}
      \item{...}
      }
}
