% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMEDIAN.R
\name{train_smedian}
\alias{train_smedian}
\title{Train seasonal median model.}
\usage{
train_smedian(.data, specials, period, ...)
}
\arguments{
\item{specials}{Specials as list defined in \code{specials_smedian}.}

\item{...}{Further arguments.}

\item{data}{Input data as \code{tsibble}.}
}
\value{
An object of class \code{SMEDIAN}.
}
\description{
Train seasonal median model (SMEDIAN).
}
