% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNAIVE2.R
\name{train_snaive2}
\alias{train_snaive2}
\title{Seasonal naive model}
\usage{
train_snaive2(.data, specials, ...)
}
\arguments{
\item{.data}{Input data as tsibble.}

\item{specials}{Specials as list defined in \code{specials_snaive2}.}

\item{...}{Currently not in use.}
}
\value{
An object of class \code{SNAIVE2}.
}
\description{
Train a seasonal naive model (SNAIVE2).
}
