% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/THETA.R
\name{train_theta}
\alias{train_theta}
\title{Train THETA model.}
\usage{
train_theta(.data, specials, ...)
}
\arguments{
\item{.data}{Input data as \code{tsibble}.}

\item{specials}{Specials as list defined in \code{specials_thetaf}.}

\item{...}{Further arguments passed to \code{forecast::thetaf()}.}
}
\value{
An object of class \code{THETA}.
}
\description{
Train a THETA model.
}
