% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{value_var}
\alias{value_var}
\title{Return value variable}
\usage{
value_var(.data)
}
\arguments{
\item{.data}{A \code{tsibble} or \code{fable}.}
}
\value{
value A character vector.
}
\description{
\code{value_var()} returns a character with the value variable, i.e.
   measured variables without "helper variables" like \code{.split}, \code{.id}, \code{.sample}, etc.
}
