% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{weather}
\alias{weather}
\title{Hourly weather data}
\format{
A time series object of class \code{tsibble} with 17.520 rows and 3 columns:
   \itemize{
      \item{\code{Time}: Date and time (index variable)}
      \item{\code{Series}: Time series name (key variable)}
      \item{\code{Value}: Measurement variable}
      }
}
\source{
\href{https://cdc.dwd.de/portal/}{CDC - Climate Data Center}
}
\usage{
data(weather)
}
\description{
Hourly \code{tsibble} with air temperature and humidity from
   the DWD ("Deutscher Wetterdienst"). The dataset contains time series data
   from 2019-01-01 00:00:00 to 2019-12-31 23:00:00 for Germany.
}
\examples{
data(weather)
}
\keyword{datasets}
