% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr.R
\name{acf_vec}
\alias{acf_vec}
\title{Estimate the sample autocorrelation of a numeric vector}
\usage{
acf_vec(x, lag_max = 24, ...)
}
\arguments{
\item{x}{Numeric vector.}

\item{lag_max}{Maximum lag as integer.}

\item{...}{Further arguments passed to \code{stats::acf()}.}
}
\value{
x Numeric vector.
}
\description{
\code{acf_vec} estimates the sample autocorrelation function
  of a numeric vector.
}
