% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{clean_data}
\alias{clean_data}
\title{Check, convert and shape the input data.}
\usage{
clean_data(data)
}
\arguments{
\item{data}{A valid tsibble, either in long or in wide format.}
}
\value{
data A valid tsibble in long format with one measurement variable.
}
\description{
This function checks whether the input data are a valid tsibble or not (regular spaced in time
   and ordered). Furthermore, implicit missing values are turned into explicit missing values (existing
   missing values are left untouched). If the data are provided in wide format, they are gathered into
   long format.
}
