% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{elec_load}
\alias{elec_load}
\title{Hourly electricity load}
\format{
A time series object of class \code{tsibble} with 35.040 rows and 4 columns:
   \itemize{
      \item{\code{Time}: Date and time (index variable)}
      \item{\code{Series}: Time series name (key variable)}
      \item{\code{BZN}: Bidding zone (key variable)}
      \item{\code{Value}: Measurement variable}
      }
}
\source{
\href{https://transparency.entsoe.eu/load-domain/r2/totalLoadR2/show}{ENTSO-E Transparency Platform}
}
\usage{
data(elec_load)
}
\description{
Hourly \code{tsibble} with electricity loads from
   the ENTSO-E Transparency Platform. The dataset contains time series data
   from 2019-01-01 00:00:00 to 2019-12-31 23:00:00 for four bidding zones
   within europe (DE, FR, SE1, NO1).
}
\examples{
data(elec_load)
}
\keyword{datasets}
