% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{elec_load}
\alias{elec_load}
\title{Hourly electricity load (actual values and forecasts)}
\format{
A time series object of class \code{tibble} with 140.160 rows and 5 columns:
   \itemize{
      \item{\code{time}: Date and time index}
      \item{\code{item}: Time series name}
      \item{\code{unit}: Measured unit}
      \item{\code{bidding_zone}: Bidding zone}
      \item{\code{value}: Measurement variable}
      }
}
\source{
\href{https://transparency.entsoe.eu/load-domain/r2/totalLoadR2/show}{ENTSO-E Transparency Platform}
}
\usage{
data(elec_load)
}
\description{
Hourly tibble with actual electricity loads and load forecasts
  from the ENTSO-E Transparency Platform. The data set contains time series data
  from 2019-01-01 00:00:00 to 2019-12-31 23:00:00 for 8 bidding zones within
  Europe (DE, DK1, ES, FI, FR, NL, NO1, SE1). The original data are on a
  quarter-hourly basis (15-minutes interval), but aggregated to hourly data.
}
\examples{
data(elec_load)
}
\keyword{datasets}
