% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{elec_load}
\alias{elec_load}
\title{Hourly electricity load and load forecasts}
\format{
A time series object of class \code{tsibble} with 122.640 rows and 5 columns:
   \itemize{
      \item{\code{Time}: Date and time (index variable)}
      \item{\code{Series}: Time series name (key variable)}
      \item{\code{Unit}: Measured unit (key variable)}
      \item{\code{BZN}: Bidding zone (key variable)}
      \item{\code{Value}: Measurement variable}
      }
}
\source{
\href{https://transparency.entsoe.eu/load-domain/r2/totalLoadR2/show}{ENTSO-E Transparency Platform}
}
\usage{
data(elec_load)
}
\description{
Hourly tsibble with electricity loads and load forecasts
  from the ENTSO-E Transparency Platform. The dataset contains time series data
  from 2019-01-01 00:00:00 to 2019-12-31 23:00:00 for 8 bidding zones within
  europe (DE, DK1, ES, FI, FR, NL, NO1, SE1). The original data are on a
  quarter-hourly basis (15-minutes interval), but aggregated to hourly data.
}
\examples{
data(elec_load)
}
\keyword{datasets}
