% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELM.R
\name{ELM}
\alias{ELM}
\title{Extreme Learning Machine (ELM)}
\usage{
ELM(formula, ...)
}
\arguments{
\item{formula}{Model specification (see "Specials" section, currently not in use...)}

\item{...}{Further arguments passed to \code{nnfor::elm()}.}
}
\value{
elm_model An object of class \code{ELM}.
}
\description{
Automatic train an Extreme Learning Machines (ELMs) model.
  This function is a wrapper for \code{nnfor::elm()}.
}
