% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_metrics.R
\name{error_metrics}
\alias{error_metrics}
\title{Estimate accuracy metrics to evaluate point forecast}
\usage{
error_metrics(fcst, test, train = NULL, period = NULL, by = "split")
}
\arguments{
\item{fcst}{A \code{fable} containing the forecasts for the models, splits, etc.}

\item{test}{A \code{tsibble} containing the testing data.}

\item{train}{A \code{tsibble} containing the training data.
If \code{train = NULL}, no MASE is calculated.}

\item{period}{Integer value. The period used for the estimation of the in-sample
MAE of seasonal naive forecast. The in-sample MAE is required for scaling the MASE.}

\item{by}{Character value. Either accuracy is estimated by \code{split} or \code{horizon}.}
}
\value{
A \code{tibble} containing the accuracy metrics for all key variables and models.
}
\description{
This function estimates several accuracy metrics to evaluate
  the accuracy of point forecasts. Either along the forecast horizon or
  along the train-test-splits. By default, the following accuracy metrics
  are provided:

   \itemize{
      \item{\code{ME}: mean error}
      \item{\code{MAE}: mean absolute error}
      \item{\code{MSE}: mean squared error}
      \item{\code{RMSE}: root mean squared error}
      \item{\code{MAPE}: mean absolute percentage error}
      \item{\code{sMAPE}: scaled mean absolute percentage error}
      \item{\code{MPE}: mean percentage error}
      \item{\code{MASE}: mean absolute scaled error}
      }
}
