% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr.R
\name{estimate_acf}
\alias{estimate_acf}
\title{Estimate the sample autocorrelation}
\usage{
estimate_acf(.data, context, lag_max = 24, ...)
}
\arguments{
\item{.data}{A \code{tibble} containing the time series data.}

\item{context}{A named \code{list} with the identifiers for \code{seried_id}, \code{value_id} and \code{index_id}.}

\item{lag_max}{Maximum lag as integer.}

\item{...}{Further arguments passed to \code{stats::acf()}.}
}
\value{
data A \code{tibble} with a column for the unique identifier of the
  time series and the columns type, lag and value.
}
\description{
The function estimates the sample autocorrelation function of a
  \code{tibble} containing several time series.
}
