% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{estimate_mode}
\alias{estimate_mode}
\title{Estimate mode of a distribution based on Kernel Density Estimation}
\usage{
estimate_mode(x, na_rm = TRUE, ...)
}
\arguments{
\item{x}{Numeric vector.}

\item{...}{Further arguments passed to \code{stats::densitiy()}.}
}
\value{
mode Numeric value. The mode of the distribution.
}
\description{
The function estimates the mode of a distribution based on Kernel Density Estimation.
}
