% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{estimate_skewness}
\alias{estimate_skewness}
\title{Estimate skewness}
\usage{
estimate_skewness(x, na_rm = TRUE)
}
\arguments{
\item{x}{Numeric vector.}

\item{na_rm}{Logical value. If \code{TRUE}, missing values are dropped.}
}
\value{
Numeric value.
}
\description{
The function estimates the skewness of a numeric vector.
}
