% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSHW.R
\name{fitted.DSHW}
\alias{fitted.DSHW}
\title{Extract fitted values from a trained DSHW model}
\usage{
\method{fitted}{DSHW}(object, ...)
}
\arguments{
\item{object}{An object of class \code{DSHW}.}

\item{...}{Currently not in use.}
}
\value{
Fitted values as tsibble.
}
\description{
Extract fitted values from a trained DSHW model.
}
