% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSHW.R
\name{fitted.DSHW}
\alias{fitted.DSHW}
\title{Extract fitted values from a DSHW.}
\usage{
\method{fitted}{DSHW}(object, ...)
}
\arguments{
\item{object}{The time series model used to produce the forecast.}

\item{...}{Further arguments.}
}
\value{

}
\description{
Extract fitted values from a DSHW.
}
