% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELM.R
\name{fitted.ELM}
\alias{fitted.ELM}
\title{Extract fitted values from a trained ELM model}
\usage{
\method{fitted}{ELM}(object, ...)
}
\arguments{
\item{object}{An object of class \code{ELM}.}

\item{...}{Currently not in use.}
}
\value{
Fitted values as tsibble.
}
\description{
Extract fitted values from a trained ELM model.
}
