% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEDIAN.R
\name{fitted.MEDIAN}
\alias{fitted.MEDIAN}
\title{Extract fitted values from a trained median model}
\usage{
\method{fitted}{MEDIAN}(object, ...)
}
\arguments{
\item{object}{An object of class \code{MEDIAN}.}

\item{...}{Currently not in use.}
}
\value{
Fitted values as tsibble.
}
\description{
Extract fitted values from a trained median model.
}
