% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLP.R
\name{fitted.MLP}
\alias{fitted.MLP}
\title{Extract fitted values from a MLP.}
\usage{
\method{fitted}{MLP}(object, ...)
}
\arguments{
\item{object}{The time series model used to produce the forecast.}

\item{...}{Further arguments.}
}
\value{
Fitted values as \code{tsibble}.
}
\description{
Extract fitted values from a MLP.
}
