% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMEAN.R
\name{fitted.SMEAN}
\alias{fitted.SMEAN}
\title{Extract fitted values from SMEAN}
\usage{
\method{fitted}{SMEAN}(object, ...)
}
\arguments{
\item{object}{The time series model used to produce the forecast.}

\item{...}{Further arguments.}
}
\value{

}
\description{
Extract fitted values from SMEAN.
}
