% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMEDIAN.R
\name{fitted.SMEDIAN}
\alias{fitted.SMEDIAN}
\title{Extract fitted values from a trained seasonal median model}
\usage{
\method{fitted}{SMEDIAN}(object, ...)
}
\arguments{
\item{object}{An object of class \code{SMEDIAN}.}

\item{...}{Currently not in use.}
}
\value{
Fitted values as tsibble.
}
\description{
Extract fitted values from a trained seasonal median model.
}
