% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMEDIAN.R
\name{fitted.SMEDIAN}
\alias{fitted.SMEDIAN}
\title{Extract fitted values from SMEDIAN.}
\usage{
\method{fitted}{SMEDIAN}(object, ...)
}
\arguments{
\item{object}{The time series model used to produce the forecast.}

\item{...}{Further arguments.}
}
\value{

}
\description{
Extract fitted values from SMEDIAN.
}
