% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNAIVE2.R
\name{fitted.SNAIVE2}
\alias{fitted.SNAIVE2}
\title{Extract fitted values from a trained seasonal naive model}
\usage{
\method{fitted}{SNAIVE2}(object, ...)
}
\arguments{
\item{object}{An object of class \code{SNAIVE2}.}

\item{...}{Currently not in use.}
}
\value{
Fitted values as tsibble.
}
\description{
Extract fitted values from a trained seasonal naive model.
}
