% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBATS.R
\name{fitted.TBATS}
\alias{fitted.TBATS}
\title{Extract fitted values from a trained TBATS model}
\usage{
\method{fitted}{TBATS}(object, ...)
}
\arguments{
\item{object}{An object of class \code{TBATS}.}

\item{...}{Currently not in use.}
}
\value{
Fitted values as tsibble.
}
\description{
Extract fitted values from a trained TBATS model.
}
