% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBATS.R
\name{fitted.TBATS}
\alias{fitted.TBATS}
\title{Extract fitted values from a TBATS.}
\usage{
\method{fitted}{TBATS}(object, ...)
}
\arguments{
\item{object}{The time series model used to produce the forecast.}

\item{...}{Further arguments.}
}
\value{

}
\description{
Extract fitted values from a TBATS.
}
