% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSHW.R
\name{forecast.DSHW}
\alias{forecast.DSHW}
\title{Forecast a trained DSHW model.}
\usage{
\method{forecast}{DSHW}(object, new_data, specials = NULL, ...)
}
\arguments{
\item{object}{The trained DSHW model used to produce the forecasts.}

\item{new_data}{Forecast horizon.}

\item{specials}{Specials are currently not in use.}

\item{...}{Additional arguments for forecast model method.}
}
\value{
An object of class "fable".
}
\description{
Forecast a trained DSHW model.
}
