% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELM.R
\name{forecast.ELM}
\alias{forecast.ELM}
\title{Forecast a trained ELM model}
\usage{
\method{forecast}{ELM}(object, new_data, specials = NULL, ...)
}
\arguments{
\item{object}{An object of class \code{ELM}.}

\item{new_data}{Forecast horizon (n-step ahead forecast)}

\item{specials}{Specials are currently not in use.}

\item{...}{Additional arguments for forecast method.}
}
\value{
An object of class \code{fable}.
}
\description{
Forecast a trained ELM model.
}
