% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEDIAN.R
\name{forecast.MEDIAN}
\alias{forecast.MEDIAN}
\title{Forecast a trained median model}
\usage{
\method{forecast}{MEDIAN}(object, new_data, specials = NULL, ...)
}
\arguments{
\item{object}{An object of class \code{MEDIAN}.}

\item{new_data}{Forecast horizon (n-step ahead forecast)}

\item{specials}{Specials are currently not in use.}

\item{...}{Additional arguments for forecast method.}
}
\value{
An object of class \code{fable}.
}
\description{
Forecast a trained median model.
}
