% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLP.R
\name{forecast.MLP}
\alias{forecast.MLP}
\title{Forecast a trained MLP model.}
\usage{
\method{forecast}{MLP}(object, new_data, specials = NULL, ...)
}
\arguments{
\item{object}{The trained MLP model used to produce the forecasts.}

\item{new_data}{Forecast horizon.}

\item{specials}{Specials are currently not in use.}

\item{...}{Additional arguments for forecast model method.}
}
\value{
An object of class "fable".
}
\description{
Forecast a trained MLP model.
}
