% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMEAN.R
\name{forecast.SMEAN}
\alias{forecast.SMEAN}
\title{Forecast a trained seasonal mean model}
\usage{
\method{forecast}{SMEAN}(object, new_data, specials = NULL, ...)
}
\arguments{
\item{object}{An object of class \code{SMEAN}.}

\item{new_data}{Forecast horizon (n-step ahead forecast)}

\item{specials}{Specials are currently not in use.}

\item{...}{Additional arguments for forecast method.}
}
\value{
An object of class \code{fable}.
}
\description{
Forecast a trained seasonal mean model.
}
