% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMEAN.R
\name{forecast.SMEAN}
\alias{forecast.SMEAN}
\title{Forecast a trained SMEAN model}
\usage{
\method{forecast}{SMEAN}(object, new_data, specials = NULL, ...)
}
\arguments{
\item{object}{Trained model.}

\item{new_data}{Forecast horizon.}

\item{specials}{Specials are currently not in use.}

\item{...}{Further arguments.}
}
\value{
A \code{fable}.
}
\description{
Forecast a trained SMEAN model.
}
