% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMEDIAN.R
\name{forecast.SMEDIAN}
\alias{forecast.SMEDIAN}
\title{Forecast a trained seasonal median model}
\usage{
\method{forecast}{SMEDIAN}(object, new_data, specials = NULL, ...)
}
\arguments{
\item{object}{An object of class \code{SMEDIAN}.}

\item{new_data}{Forecast horizon (n-step ahead forecast)}

\item{specials}{Specials are currently not in use.}

\item{...}{Additional arguments for forecast method.}
}
\value{
An object of class \code{fable}.
}
\description{
Forecast a trained seasonal median model.
}
