% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMEDIAN.R
\name{forecast.SMEDIAN}
\alias{forecast.SMEDIAN}
\title{Forecast a trained SMEDIAN model.}
\usage{
\method{forecast}{SMEDIAN}(object, new_data, specials = NULL, ...)
}
\arguments{
\item{object}{Trained model.}

\item{new_data}{Forecast horizon.}

\item{specials}{Specials are currently not in use.}

\item{...}{Further arguments.}
}
\value{
A \code{fable}.
}
\description{
Forecast a trained SMEDIAN model.
}
