% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNAIVE2.R
\name{forecast.SNAIVE2}
\alias{forecast.SNAIVE2}
\title{Forecast a trained seasonal naive model}
\usage{
\method{forecast}{SNAIVE2}(object, new_data, specials = NULL, ...)
}
\arguments{
\item{object}{An object of class \code{SNAIVE2}.}

\item{new_data}{Forecast horizon (n-step ahead forecast)}

\item{specials}{Specials are currently not in use.}

\item{...}{Additional arguments for forecast method.}
}
\value{
An object of class \code{fable}.
}
\description{
Forecast a trained seasonal naive model.
}
