% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBATS.R
\name{forecast.TBATS}
\alias{forecast.TBATS}
\title{Forecast a trained TBATS model}
\usage{
\method{forecast}{TBATS}(object, new_data, specials = NULL, ...)
}
\arguments{
\item{object}{An object of class \code{TBATS}.}

\item{new_data}{Forecast horizon (n-step ahead forecast)}

\item{specials}{Specials are currently not in use.}

\item{...}{Additional arguments for forecast method.}
}
\value{
An object of class \code{fable}.
}
\description{
Forecast a trained TBATS model.
}
