% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBATS.R
\name{forecast.TBATS}
\alias{forecast.TBATS}
\title{Forecast a trained TBATS model}
\usage{
\method{forecast}{TBATS}(object, new_data, specials = NULL, ...)
}
\arguments{
\item{object}{The trained TBATS model used to produce the forecasts.}

\item{new_data}{Forecast horizon.}

\item{specials}{Specials are currently not in use.}

\item{...}{Additional arguments for forecast model method.}
}
\value{
An object of class "fable".
}
\description{
Forecast a trained TBATS model.
}
