% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_split.R
\name{initialize_split}
\alias{initialize_split}
\title{Initialize a plan for train-test split}
\usage{
initialize_split(
  main_frame,
  context,
  type = c("first", "last", "prob"),
  value = NULL
)
}
\arguments{
\item{main_frame}{A \code{tibble} containing the time series data.}

\item{context}{A named \code{list} with the identifiers for \code{seried_id}, \code{value_id} and \code{index_id}.}

\item{type}{The type for the initial split. Possible values are \code{"first"}, \code{"last"}, \code{"prob"}.}

\item{value}{Numeric value specifying the split.}
}
\value{
A \code{tibble}
}
\description{
The function creates the initial split into training and
  testing data.
}
