% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate_missing.R
\name{interpolate_missing}
\alias{interpolate_missing}
\title{Interpolate missing values}
\usage{
interpolate_missing(x, periods, ...)
}
\arguments{
\item{x}{Numeric vector.}

\item{periods}{Numeric vector. The seasonal periods of the time series.}

\item{...}{Further arguments passed to \code{forecast::msts()} or \code{forecast::na.interp()}.}
}
\value{
Numeric vector.
}
\description{
The function \code{interpolate_missing()} is a wrapper
  for \code{forecast::na.interp()}, working with numeric vectors. For
  non-seasonal time series, linear interpolation is used and for
  seasonal time series, the series is decomposed via STL and the
  seasonally adjusted series is linearly interpolated and the seasonal
  component is added back.
}
