% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{log_time}
\alias{log_time}
\title{Create string with elapsed time.}
\usage{
log_time(text, start = Sys.time(), end = Sys.time(), digits = 1, ...)
}
\arguments{
\item{text}{Character value. The text for the header.}

\item{start}{Start date and time.}

\item{end}{End date and time.}

\item{digits}{Integer value. The number of digits for rounding.}
}
\value{
Character value.
}
\description{
Create a \code{glue} object with the elapsed time between start
  and end.
}
