% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{mae_vec}
\alias{mae_vec}
\title{Calculate the mean absolute error}
\usage{
mae_vec(truth, estimate, na_rm = TRUE)
}
\arguments{
\item{truth}{Numeric vector containing the actual values.}

\item{estimate}{Numeric vector containing the forecasts.}

\item{na_rm}{Logical value. If \code{TRUE}, missing values are removed.}
}
\value{
Numeric value.
}
\description{
Calculate the mean absolute error of a numeric vector.
   \code{mae()} is a metric that is in the same units as the original data.
}
