% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_errors.R
\name{make_errors}
\alias{make_errors}
\title{Calculate forecast errors}
\usage{
make_errors(future_frame, main_frame, context)
}
\arguments{
\item{future_frame}{A \code{tibble} containing the forecasts for the models, splits, etc.}

\item{main_frame}{A \code{tibble} containing the actual values.}

\item{series_id}{Character value specifying the series identifier.}

\item{value_id}{Character value specifying the value identifier.}

\item{index_id}{Character value specifying the index identifier.}
}
\value{
errors A \code{tibble} with forecast errors and
  percentage forecast errors.
}
\description{
\code{errors} calculates the forecast errors (error) and
  percentage forecast errors (pct_error).
}
