% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_split.R
\name{make_split}
\alias{make_split}
\title{...}
\usage{
make_split(
  main,
  context,
  type,
  value,
  n_ahead,
  n_skip = 0,
  n_lag = 0,
  mode = "slide",
  exceed = TRUE
)
}
\arguments{
\item{context}{...}

\item{type}{...}

\item{value}{...}

\item{n_ahead}{...}

\item{n_skip}{...}

\item{n_lag}{...}

\item{mode}{...}

\item{exceed}{...}

\item{data}{...}
}
\value{

}
\description{
...
}
