% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_tsibble.R
\name{make_tsibble}
\alias{make_tsibble}
\title{Convert tibble to tsibble}
\usage{
make_tsibble(main_frame, context)
}
\arguments{
\item{main_frame}{A \code{tibble} containing the time series data.}

\item{context}{A named \code{list} with the identifiers for \code{seried_id}, \code{value_id} and \code{index_id}.}
}
\value{
main_frame Same structure as before, just stored as \code{tsibble}.
}
\description{
\code{make_tsibble} is a helper function to convert a
  \code{tibble} to a \code{tsibble}.
}
