% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{mape_vec}
\alias{mape_vec}
\title{Calculate the mean absolute percentage error}
\usage{
mape_vec(truth, estimate, na_rm = TRUE)
}
\arguments{
\item{truth}{Numeric vector containing the actual values.}

\item{estimate}{Numeric vector containing the forecasts.}

\item{na_rm}{Logical value. If \code{TRUE}, missing values are removed.}
}
\value{
Numeric value.
}
\description{
Calculate the mean absolute percentage error of a numeric vector.
   \code{mape()} is a metric that is in relative units.
}
