% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{me}
\alias{me}
\title{Calculate the mean error}
\usage{
me(data, ...)
}
\arguments{
\item{data}{A \code{data.frame}, \code{tibble} or \code{tsibble} containing
the \code{truth} and \code{estimate}.}

\item{...}{Not currently used.}
}
\value{

}
\description{
Calculate the mean error. \code{me()} is a metric that is in
   the same units as the original data.
}
