% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEDIAN.R
\name{MEDIAN}
\alias{MEDIAN}
\title{Median model}
\usage{
MEDIAN(formula, ...)
}
\arguments{
\item{formula}{Model specification (see "Specials" section, currently not in use ...).}

\item{...}{Further arguments.}
}
\value{
median_model An object of class \code{MEDIAN}.
}
\description{
Automatic train a median model (MEDIAN).
}
