% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSHW.R
\name{model_sum.DSHW}
\alias{model_sum.DSHW}
\title{Provide a succinct summary of the DSHW}
\usage{
\method{model_sum}{DSHW}(x)
}
\arguments{
\item{x}{The DSHW to summarize.}
}
\value{

}
\description{
Provide a succinct summary of the DSHW.
}
