% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSHW.R
\name{model_sum.DSHW}
\alias{model_sum.DSHW}
\title{Provide a succinct summary of a trained DSHW model}
\usage{
\method{model_sum}{DSHW}(object)
}
\arguments{
\item{object}{An object of class \code{DSHW}.}
}
\value{
Model summary as character value.
}
\description{
Provide a succinct summary of a trained DSHW model.
}
