% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELM.R
\name{model_sum.ELM}
\alias{model_sum.ELM}
\title{Provide a succinct summary of the ELM.}
\usage{
\method{model_sum}{ELM}(x)
}
\arguments{
\item{object}{The ELM to summarize.}
}
\value{

}
\description{
Provide a succinct summary of the ELM.
}
