% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEDIAN.R
\name{model_sum.MEDIAN}
\alias{model_sum.MEDIAN}
\title{Provide a succinct summary of a trained median model}
\usage{
\method{model_sum}{MEDIAN}(x)
}
\arguments{
\item{object}{An object of class \code{MEDIAN}.}
}
\value{
Model summary as character value.
}
\description{
Provide a succinct summary of a trained median model.
}
