% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLP.R
\name{model_sum.MLP}
\alias{model_sum.MLP}
\title{Provide a succinct summary of the MLP}
\usage{
\method{model_sum}{MLP}(x)
}
\arguments{
\item{x}{The MLP to summarize.}
}
\value{
Model summary as character value.
}
\description{
Provide a succinct summary of the MLP.
}
