% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMEAN.R
\name{model_sum.SMEAN}
\alias{model_sum.SMEAN}
\title{Provide a succinct summary of the SMEAN}
\usage{
\method{model_sum}{SMEAN}(x)
}
\arguments{
\item{x}{The SMEAN to summarize.}
}
\value{
Model summary as character value.
}
\description{
Provide a succinct summary of the SMEAN.
}
