% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMEDIAN.R
\name{model_sum.SMEDIAN}
\alias{model_sum.SMEDIAN}
\title{Provide a succinct summary of a trained seasonal median model}
\usage{
\method{model_sum}{SMEDIAN}(object)
}
\arguments{
\item{object}{An object of class \code{SMEDIAN}.}
}
\value{
Model summary as character value.
}
\description{
Provide a succinct summary of a trained seasonal median model.
}
