% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMEDIAN.R
\name{model_sum.SMEDIAN}
\alias{model_sum.SMEDIAN}
\title{Provide a succinct summary of the SMEDIAN}
\usage{
\method{model_sum}{SMEDIAN}(x)
}
\arguments{
\item{x}{The SMEDIAN to summarize.}
}
\value{

}
\description{
Provide a succinct summary of the SMEDIAN.
}
