% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNAIVE2.R
\name{model_sum.SNAIVE2}
\alias{model_sum.SNAIVE2}
\title{Provide a succinct summary of a trained seasonal naive model}
\usage{
\method{model_sum}{SNAIVE2}(object)
}
\arguments{
\item{object}{An object of class \code{SNAIVE2}.}
}
\value{
Model summary as character value.
}
\description{
Provide a succinct summary of a trained seasonal naive model.
}
