% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{mpe_vec}
\alias{mpe_vec}
\title{Calculate the mean percentage error}
\usage{
mpe_vec(truth, estimate, na_rm = TRUE, ...)
}
\arguments{
\item{truth}{Numeric vector containing the actual values.}

\item{estimate}{Numeric vector containing the forecasts.}

\item{na_rm}{Logical value. If \code{TRUE}, missing values are removed.}

\item{...}{Further arguments passed to \code{metric_vec_template()}.}
}
\value{
Numeric value.
}
\description{
Calculate the mean percentage error of a numeric vector.
   \code{mpe()} is a metric that is in relative units.
}
