% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{mse}
\alias{mse}
\title{Calculate the mean squared error}
\usage{
mse(data, ...)
}
\arguments{
\item{data}{A \code{data.frame}, \code{tibble} or \code{tsibble} containing
the \code{truth} and \code{estimate}.}

\item{...}{Not currently used.}
}
\value{

}
\description{
Calculate the mean squared error. \code{mse()} is a metric
   that is in quadratic units.
}
