% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{mse_vec}
\alias{mse_vec}
\title{Calculate the mean squared error}
\usage{
mse_vec(truth, estimate, na_rm = TRUE, ...)
}
\arguments{
\item{truth}{Numeric vector containing the actual values.}

\item{estimate}{Numeric vector containing the forecasts.}

\item{na_rm}{Logical value. If \code{TRUE}, missing values are removed.}

\item{...}{Further arguments passed to \code{metric_vec_template()}.}
}
\value{
Numeric value.
}
\description{
Calculate the mean squared error of a numeric vector.
   \code{mse()} is a metric that is in quadratic units.
}
