% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_density.R
\name{plot_density}
\alias{plot_density}
\title{Plot the density via Kernel Density Estimator.}
\usage{
plot_density(
  data,
  title = NULL,
  subtitle = NULL,
  xlab = NULL,
  ylab = NULL,
  caption = NULL,
  line_width = 0.1,
  line_type = "solid",
  line_color = "#31688EFF",
  line_fill = "#31688EFF",
  line_alpha = 0.5,
  theme_set = theme_tscv(),
  theme_config = list()
)
}
\arguments{
\item{data}{A valid tsibble in long format with one measurement variable.}

\item{title}{Title of the plot.}

\item{subtitle}{Subtitle of the plot.}

\item{xlab}{Label for the x-axis.}

\item{ylab}{Label for the y-axis.}

\item{caption}{Caption of the plot.}

\item{line_width}{Numeric value defining the line width of the kernel density estimator.}

\item{line_type}{Integer value defining the line type of the kernel density estimator.}

\item{line_color}{Character value defining the line color of the kernel density estimator.}

\item{line_fill}{Character value defining the fill color for the area under the kernel density estimator.}

\item{line_alpha}{Numeric value defining the transparency of the area under the kernel density estimator.}

\item{theme_set}{A complete ggplot2 theme.}

\item{theme_config}{A list with further arguments passed to \code{ggplot2::theme()}.}
}
\value{
p An object of class ggplot.
}
\description{
Plot the density of one or more time series via Kernel Density Estimator.
}
