% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_error_metrics.R
\name{plot_error_metrics}
\alias{plot_error_metrics}
\title{Plot forecast accuracy metrics}
\usage{
plot_error_metrics(
  data,
  model = NULL,
  metric = "MASE",
  title = NULL,
  subtitle = NULL,
  ylab = NULL,
  xlab = NULL,
  caption = NULL,
  line_width = 1,
  line_type = "solid",
  point_size = 0,
  point_shape = 19,
  point_alpha = 1,
  theme_set = theme_tscv(),
  theme_config = list(),
  ...
)
}
\arguments{
\item{data}{A \code{tibble} containing the accuracy metrics, i.e. the result of a call to \code{error_metrics()}.}

\item{model}{Character vector defining the forecasting models.}

\item{metric}{Character vector defining the accuracy measures.}

\item{title}{Title for the plot.}

\item{subtitle}{Subtitle for the plot.}

\item{ylab}{Label for the y-axis.}

\item{xlab}{Label for the x-axis.}

\item{caption}{Caption (footnote) for the plot.}

\item{line_width}{Numeric value. Line width.}

\item{line_type}{Character value defining the line type.}

\item{point_size}{Numeric value. Point size.}

\item{point_shape}{Integer value. Point shape.}

\item{point_alpha}{Numeric value between 0 and 1. The transparency of the points.}

\item{theme_set}{A complete ggplot2 theme.}

\item{theme_config}{A list with further arguments passed to \code{ggplot2::theme()}.}

\item{...}{Further arguments passed to \code{ggplot2::geom_line()} or \code{ggplot2::geom_point()}.}
}
\value{
p An object of class ggplot.
}
\description{
Plot forecast accuracy metrics, either along the forecast horizon or the slices. The user can define
   the forecasting models and accuracy metrics.
}
