% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_forecast.R
\name{plot_forecast}
\alias{plot_forecast}
\title{Plot the actual and fitted values and forecasts.}
\usage{
plot_forecast(
  fcst,
  data,
  split = NULL,
  model = NULL,
  include = 24,
  title = NULL,
  subtitle = NULL,
  xlab = NULL,
  ylab = NULL,
  caption = NULL,
  line_width = 1,
  point_size = 2,
  point_alpha = 0,
  theme_set = theme_tscv(),
  theme_config = list()
)
}
\arguments{
\item{fcst}{A `fable` containing the forecasts for the models, splits, etc.}

\item{data}{A `tsibble` containing the training and testing data.}

\item{split}{Integer vector. The split id (time slices) of the time series cross-validation.}

\item{model}{Character vector. The forecasting model(s).}

\item{include}{Integer value. The number of actual values (training) to be included.}

\item{title}{Title of the plot.}

\item{subtitle}{Subtitle of the plot.}

\item{xlab}{Label for the x-axis.}

\item{ylab}{Label for the y-axis.}

\item{caption}{Caption of the plot.}

\item{line_width}{Numeric value defining the line width.}

\item{point_size}{Numeric value defining the point size.}

\item{point_alpha}{Numeric value. The transparency of the points.}

\item{theme_set}{A complete ggplot2 theme.}

\item{theme_config}{A list with further arguments passed to \code{ggplot2::theme()}.}
}
\value{
p An object of class ggplot.
}
\description{
Plot the actual and fitted values and forecasts for specific target variable(s), split(s) and forecasting method(s).
}
