% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_forecast.R
\name{plot_forecast}
\alias{plot_forecast}
\title{Plot the actual and fitted values and forecasts}
\usage{
plot_forecast(
  data,
  variable,
  slice,
  model,
  include = 24,
  title = "Actual values and forecast",
  subtitle = NULL,
  xlab = "Date",
  ylab = NULL,
  caption = NULL,
  line_width = 1,
  point_size = 2,
  point_alpha = 0,
  base_size = 11
)
}
\arguments{
\item{data}{A tsibble according to the adice data model}

\item{variable}{Character value defining the target variable}

\item{slice}{Character value defining the time slice of the rolling forecast}

\item{model}{Character vector defining the forecasting methods}

\item{include}{Integer value, the number of actual values (training) to be included}

\item{title}{Title of the plot (default select.target)}

\item{xlab}{Label for the x-axis}

\item{ylab}{Label for the y-axis}

\item{caption}{Caption of the plot}

\item{line_width}{Numeric value defining the line width}

\item{point_size}{Numeric value defining the point size}

\item{point_alpha}{Transparency of the points (default 1, no transparency)}

\item{base_size}{Integer value. Base font size}

\item{add_fitted}{Logical value if fitted values are plotted (TRUE) or not (default FALSE)}
}
\value{
p An object of class ggplot
}
\description{
Plot the actual and fitted values and forecasts for specific target variable(s), slice(s) and forecasting method(s).
}
