% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_histogram.R
\name{plot_histogram}
\alias{plot_histogram}
\title{Plot the histogram of time series.}
\usage{
plot_histogram(
  .data,
  x,
  facet = NULL,
  color = NULL,
  fill = NULL,
  title = NULL,
  subtitle = NULL,
  xlab = NULL,
  ylab = NULL,
  caption = NULL,
  line_color = "grey35",
  line_width = 0.5,
  fill_color = "grey35",
  fill_alpha = 1,
  theme_set = theme_tscv(),
  theme_config = list(),
  ...
)
}
\arguments{
\item{title}{Title of the plot.}

\item{subtitle}{Subtitle of the plot.}

\item{xlab}{Label for the x-axis.}

\item{ylab}{Label for the y-axis.}

\item{caption}{Caption of the plot.}

\item{theme_set}{A complete ggplot2 theme.}

\item{theme_config}{A list with further arguments passed to \code{ggplot2::theme()}.}

\item{...}{Further arguments passed to \code{ggplot2::geom_histogram()}.}

\item{data}{A valid tsibble in long format with one measurement variable.}

\item{bar_line_color}{Character value defining the outline color of the histrogram bars.}

\item{bar_fill_color}{Character value defining the color of the histrogram bars.}

\item{bar_alpha}{Numeric value defining the transparency of the histogram bars.}
}
\value{
p An object of class ggplot.
}
\description{
Plot the historgram of one or more time series.
}
