% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_point.R
\name{plot_point}
\alias{plot_point}
\title{Plot data as scatterplot}
\usage{
plot_point(
  data,
  x,
  y,
  facet_var = NULL,
  facet_scale = "free",
  color = NULL,
  title = NULL,
  subtitle = NULL,
  xlab = NULL,
  ylab = NULL,
  caption = NULL,
  point_size = 1.5,
  point_type = 16,
  point_color = "grey35",
  point_alpha = 1,
  theme_set = theme_tscv(),
  theme_config = list(),
  ...
)
}
\arguments{
\item{data}{A \code{data.frame}, \code{tibble} or \code{tsibble} in long format.}

\item{x}{Unquoted column within \code{.data}.}

\item{y}{Unquoted column within \code{.data} containing numeric values.}

\item{facet_var}{Unquoted column within \code{.data} (facet).}

\item{facet_scale}{Character value defining axis scaling (\code{facet_var = "free"} or \code{facet_var = "fixed"}).}

\item{color}{Unquoted column within \code{.data} (color).}

\item{title}{Title of the plot.}

\item{subtitle}{Subtitle of the plot.}

\item{xlab}{Label for the x-axis.}

\item{ylab}{Label for the y-axis.}

\item{caption}{Caption of the plot.}

\item{point_size}{Numeric value defining the point size.}

\item{point_type}{Numeric value defining the point type.}

\item{point_color}{Character value defining the point color (ignored if \code{color} is present).}

\item{point_alpha}{Numeric value defining the transparency of the points.}

\item{theme_set}{A complete ggplot2 theme.}

\item{theme_config}{A list with further arguments passed to \code{ggplot2::theme()}.}

\item{...}{Currently not in use.}
}
\value{
p An object of class ggplot.
}
\description{
Plot one or more variables as scatterplot.
}
