% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_qq.R
\name{plot_qq}
\alias{plot_qq}
\title{Quantile-Quantile plot (normal)}
\usage{
plot_qq(
  data,
  title = NULL,
  subtitle = NULL,
  xlab = "Theoretical quantile",
  ylab = "Sample quantile",
  caption = NULL,
  point_size = 4,
  point_shape = 16,
  point_color = "#31688EFF",
  point_fill = "#31688EFF",
  point_alpha = 0.5,
  line_width = 0.25,
  line_type = "solid",
  line_color = "grey35"
)
}
\arguments{
\item{data}{A tsibble containing the columns time, variable and value.}

\item{title}{Title of the plot (default select.target)}

\item{subtitle}{Subtitle of the plot (default NULL)}

\item{xlab}{Label for the x-axis}

\item{ylab}{Label for the y-axis}

\item{caption}{Caption of the plot}

\item{point.size}{Numeric value defining the point size}

\item{point.shape}{Integer value defining the point shape}

\item{point.color}{Character value defining the point color}

\item{point.alpha}{Numeric value defining the transperency of points}

\item{line.width}{Numeric value defining the line width (45-degree line)}

\item{line.type}{Integer value defining the line type (45-degree line)}

\item{line.color}{Character value defining the line color (45-degree line)}
}
\value{
p An object of class ggplot
}
\description{
Quantile-Quantile plot (normal) of one or more time series.
}
