% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_qq.R
\name{plot_qq}
\alias{plot_qq}
\title{Quantile-Quantile plot}
\usage{
plot_qq(
  .data,
  x,
  facet = NULL,
  color = NULL,
  title = NULL,
  subtitle = NULL,
  xlab = NULL,
  ylab = NULL,
  caption = NULL,
  point_size = 2,
  point_shape = 16,
  point_color = "grey35",
  point_fill = "grey35",
  point_alpha = 0.25,
  line_width = 0.25,
  line_type = "solid",
  line_color = "grey35",
  line_alpha = 1,
  band_color = "grey35",
  band_alpha = 0.25,
  theme_set = theme_tscv(),
  theme_config = list(),
  ...
)
}
\arguments{
\item{.data}{A \code{data.frame}, \code{tibble} or \code{tsibble} in long format.}

\item{x}{Unquoted column within \code{.data} containing numeric values.}

\item{facet}{Unquoted column within \code{.data} (facet).}

\item{color}{Unquoted column within \code{.data} (color).}

\item{title}{Title of the plot.}

\item{subtitle}{Subtitle of the plot.}

\item{xlab}{Label for the x-axis.}

\item{ylab}{Label for the y-axis.}

\item{caption}{Caption of the plot.}

\item{point_size}{Numeric value defining the point size.}

\item{point_shape}{Integer value defining the point shape.}

\item{point_color}{Character value defining the point color (ignored if \code{color} is present).}

\item{point_fill}{Character value defining the fill color (ignored if \code{color} is present).}

\item{point_alpha}{Numeric value defining the transparency of points.}

\item{line_width}{Numeric value defining the line width (45-degree line).}

\item{line_type}{Integer value defining the line type (45-degree line).}

\item{line_color}{Character value defining the line color (45-degree line).}

\item{line_alpha}{Numeric value defining the transparency of the line.}

\item{band_color}{Character value defining the fill color of the confidence bands.}

\item{band_alpha}{Numeric value defining the transparency of the confidence bands.}

\item{theme_set}{A complete ggplot2 theme.}

\item{theme_config}{A list with further arguments passed to \code{ggplot2::theme()}.}

\item{...}{Further arguments passed to \code{qqplotr::stat_qq_point()},
\code{qqplotr::stat_qq_line()}, \code{qqplotr::stat_qq_band()}.}
}
\value{
p An object of class ggplot.
}
\description{
Quantile-Quantile plot of one or more time series.
}
