% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_qq.R
\name{plot_qq}
\alias{plot_qq}
\title{Quantile-Quantile plot (normal).}
\usage{
plot_qq(
  data,
  title = NULL,
  subtitle = NULL,
  xlab = NULL,
  ylab = NULL,
  caption = NULL,
  point_size = 2,
  point_shape = 16,
  point_color = "#31688EFF",
  point_fill = "#31688EFF",
  point_alpha = 0.25,
  line_width = 0.25,
  line_type = "solid",
  line_color = "grey35",
  theme_set = theme_tscv(),
  theme_config = list()
)
}
\arguments{
\item{data}{A valid tsibble in long format with one measurement variable.}

\item{title}{Title of the plot (default select.target)}

\item{subtitle}{Subtitle of the plot (default NULL)}

\item{xlab}{Label for the x-axis}

\item{ylab}{Label for the y-axis}

\item{caption}{Caption of the plot}

\item{point_size}{Numeric value defining the point size}

\item{point_shape}{Integer value defining the point shape}

\item{point_color}{Character value defining the point color}

\item{point_alpha}{Numeric value defining the transperency of points}

\item{line_width}{Numeric value defining the line width (45-degree line)}

\item{line_type}{Integer value defining the line type (45-degree line)}

\item{line_color}{Character value defining the line color (45-degree line)}

\item{theme_set}{A complete ggplot2 theme.}

\item{theme_config}{A list with further arguments passed to \code{ggplot2::theme()}.}
}
\value{
p An object of class ggplot
}
\description{
Quantile-Quantile plot (normal) of one or more time series.
}
