% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_season.R
\name{plot_season}
\alias{plot_season}
\title{Seasonal plot of a time series.}
\usage{
plot_season(
  data,
  title = NULL,
  subtitle = NULL,
  xlab = NULL,
  ylab = NULL,
  caption = NULL,
  line_width = 0.75,
  line_type = "solid",
  line_color = "#31688EFF",
  line_alpha = 1,
  theme_set = theme_tscv(),
  theme_config = list()
)
}
\arguments{
\item{data}{A tsibble containing the columns time, variable and value.}

\item{title}{Title of the plot.}

\item{subtitle}{Subtitle of the plot.}

\item{xlab}{Label for the x-axis.}

\item{ylab}{Label for the y-axis.}

\item{caption}{Caption of the plot.}

\item{line_width}{Numeric value defining the line width (time series).}

\item{line_type}{Numeric value defining the line type (time series).}

\item{line_color}{Character value defining the line color (time series).}

\item{line_alpha}{Numeric value defining the transparency of the line.}

\item{theme_set}{A complete ggplot2 theme.}

\item{theme_config}{A list with further arguments passed to \code{ggplot2::theme()}.}
}
\value{
p An object of class ggplot.
}
\description{
Seasonal plot of one or more time series.
}
