% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSHW.R
\name{residuals.DSHW}
\alias{residuals.DSHW}
\title{Extract residuals from a trained DSHW model}
\usage{
\method{residuals}{DSHW}(object, ...)
}
\arguments{
\item{object}{An object of class \code{DSHW}.}

\item{...}{Currently not in use.}
}
\value{
Residuals as tsibble.
}
\description{
Extract residuals from a trained DSHW model.
}
