% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSHW.R
\name{residuals.DSHW}
\alias{residuals.DSHW}
\title{Extract residuals from a DSHW}
\usage{
\method{residuals}{DSHW}(object, ...)
}
\arguments{
\item{object}{The time series model used to produce the forecast.}

\item{...}{Further arguments.}
}
\value{

}
\description{
Extract residuals from a DSHW.
}
