% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELM.R
\name{residuals.ELM}
\alias{residuals.ELM}
\title{Extract residuals from a ELM.}
\usage{
\method{residuals}{ELM}(object, ...)
}
\arguments{
\item{object}{The time series model used to produce the forecast.}

\item{...}{Further arguments.}
}
\value{

}
\description{
Extract residuals from a ELM.
}
