% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEDIAN.R
\name{residuals.MEDIAN}
\alias{residuals.MEDIAN}
\title{Extract residuals from a trained median model}
\usage{
\method{residuals}{MEDIAN}(object, ...)
}
\arguments{
\item{object}{An object of class \code{MEDIAN}.}

\item{...}{Currently not in use.}
}
\value{
Fitted values as tsibble.
}
\description{
Extract residuals from a trained median model.
}
