% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLP.R
\name{residuals.MLP}
\alias{residuals.MLP}
\title{Extract residuals from a MLP}
\usage{
\method{residuals}{MLP}(object, ...)
}
\arguments{
\item{object}{The time series model used to produce the forecast.}

\item{...}{Further arguments.}
}
\value{

}
\description{
Extract residuals from a MLP.
}
