% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMEAN.R
\name{residuals.SMEAN}
\alias{residuals.SMEAN}
\title{Extract residuals from a trained seasonal mean model}
\usage{
\method{residuals}{SMEAN}(object, ...)
}
\arguments{
\item{object}{An object of class \code{SMEAN}.}

\item{...}{Currently not in use.}
}
\value{
Fitted values as tsibble.
}
\description{
Extract residuals from a trained seasonal mean model.
}
