% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMEAN.R
\name{residuals.SMEAN}
\alias{residuals.SMEAN}
\title{Extract residuals from SMEAN}
\usage{
\method{residuals}{SMEAN}(object, ...)
}
\arguments{
\item{object}{The time series model used to produce the forecast.}

\item{...}{Further arguments.}
}
\value{

}
\description{
Extract residuals from SMEAN.
}
