% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMEDIAN.R
\name{residuals.SMEDIAN}
\alias{residuals.SMEDIAN}
\title{Extract residuals from a trained seasonal median model}
\usage{
\method{residuals}{SMEDIAN}(object, ...)
}
\arguments{
\item{object}{An object of class \code{SMEDIAN}.}

\item{...}{Currently not in use.}
}
\value{
Fitted values as tsibble.
}
\description{
Extract residuals from a trained seasonal median model.
}
