% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMEDIAN.R
\name{residuals.SMEDIAN}
\alias{residuals.SMEDIAN}
\title{Extract residuals from SMEDIAN}
\usage{
\method{residuals}{SMEDIAN}(object, ...)
}
\arguments{
\item{object}{The time series model used to produce the forecast.}

\item{...}{Further arguments.}
}
\value{

}
\description{
Extract residuals from SMEDIAN.
}
