% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNAIVE2.R
\name{residuals.SNAIVE2}
\alias{residuals.SNAIVE2}
\title{Extract residuals from a trained seasonal naive model}
\usage{
\method{residuals}{SNAIVE2}(object, ...)
}
\arguments{
\item{object}{An object of class \code{SNAIVE2}.}

\item{...}{Currently not in use.}
}
\value{
Fitted values as tsibble.
}
\description{
Extract residuals from a trained seasonal naive model.
}
