% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBATS.R
\name{residuals.TBATS}
\alias{residuals.TBATS}
\title{Extract residuals from a trained TBATS model}
\usage{
\method{residuals}{TBATS}(object, ...)
}
\arguments{
\item{object}{An object of class \code{TBATS}.}

\item{...}{Currently not in use.}
}
\value{
Fitted values as tsibble.
}
\description{
Extract residuals from a trained TBATS model.
}
