% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBATS.R
\name{residuals.TBATS}
\alias{residuals.TBATS}
\title{Extract residuals from a TBATS}
\usage{
\method{residuals}{TBATS}(object, ...)
}
\arguments{
\item{object}{The time series model used to produce the forecast.}

\item{...}{Further arguments.}
}
\value{

}
\description{
Extract residuals from a TBATS.
}
