% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{rmse_vec}
\alias{rmse_vec}
\title{Calculate the root mean squared error}
\usage{
rmse_vec(truth, estimate, na_rm = TRUE)
}
\arguments{
\item{truth}{Numeric vector containing the actual values.}

\item{estimate}{Numeric vector containing the forecasts.}

\item{na_rm}{Logical value. If \code{TRUE}, missing values are removed.}
}
\value{
Numeric value.
}
\description{
Calculate the root mean squared error of a numeric vector.
   \code{rmse()} is a metric that is in the same units as the original data.
}
