% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors_tscv.R
\name{scale_color_tscv}
\alias{scale_color_tscv}
\title{Color scale constructor for tscv colors.}
\usage{
scale_color_tscv(palette = "main", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character value. The name of the palette in tscv_palettes.}

\item{discrete}{Logical value indicating whether color aesthetic is discrete or not.}

\item{reverse}{Logical value indicating whether the palette should be reversed.}

\item{...}{Additional arguments passed to \code{discrete_scale()} or
\code{scale_color_gradientn()}.}
}
\description{
Color scale constructor for tscv colors.
}
