% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMEAN.R
\name{SMEAN}
\alias{SMEAN}
\title{Seasonal mean model}
\usage{
SMEAN(formula, ...)
}
\arguments{
\item{formula}{Model specification (see "Specials" section, currently not in use ...).}

\item{...}{Further arguments.}
}
\value{
smean_model An object of class \code{SMEAN}.
}
\description{
Automatic train a seasonal mean model (SMEAN). This is
  equivalent to a linear regression against seasonal dummy variables only,
  i.e. \code{TSLM(value ~ season())}.
}
