% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMEDIAN.R
\name{SMEDIAN}
\alias{SMEDIAN}
\title{Seasonal median model}
\usage{
SMEDIAN(formula, ...)
}
\arguments{
\item{formula}{Model specification (see "Specials" section, currently not in use ...).}

\item{...}{Further arguments.}
}
\value{
smedian_model An object of class \code{SMEDIAN}.
}
\description{
Automatic train a seasonal median model (SMEDIAN).
}
