% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{smooth_outlier}
\alias{smooth_outlier}
\title{Identify and replace outliers}
\usage{
smooth_outlier(x, period, ...)
}
\arguments{
\item{x}{Numeric vector.}

\item{period}{Numeric vector. The seasonal periods of the time series.}

\item{...}{Further arguments passed to \code{forecast::msts()} or \code{forecast::tsoutliers()}.}
}
\value{
Numeric vector.
}
\description{
The function \code{smooth_outlier()} is a wrapper
  for \code{forecast::tsoutliers()}, working with numeric vectors. For
  non-seasonal time series, the supsmu method is used. For seasonal
  time series, the series is decomposed via STL and the IQR method is
  used on the remainder component. Values outside the range are
  linear interpolated on the remainder and the series is reconstructed
  with the corrected remainder component.
}
