% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_outlier.R
\name{smooth_outlier}
\alias{smooth_outlier}
\title{Identify and smooth outliers.}
\usage{
smooth_outlier(data, alpha = 0.05)
}
\arguments{
\item{data}{A \code{tsibble} containing the time series data.}

\item{alpha}{Numeric value. Controls the width of the limits.}
}
\value{
data A \code{tsibble} containing the outlier adjusted time series.
}
\description{
\code{smooth_outliers} identifies and adjusts outliers. The time
   series is decomposed via \code{STL} decomposition into trend, season and
   remainder. The IQR method is then applied to the remainder to
   identify and replace outliers with NAs. Afterwards, the NAs are linearly
   interpolated and the time series is additive reconstructed.
}
