% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNAIVE2.R
\name{SNAIVE2}
\alias{SNAIVE2}
\title{Seasonal naive model}
\usage{
SNAIVE2(formula, ...)
}
\arguments{
\item{formula}{Model specification (see "Specials" section, currently not in use ...).}

\item{...}{Further arguments.}
}
\value{
snaive2_model An object of class \code{SNAIVE2}.
}
\description{
Automatic train a seasonal naive model (SNAIVE2)..
}
