% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_split.R
\name{split_index}
\alias{split_index}
\title{Create indices for train and test splits.}
\usage{
split_index(
  n_total,
  n_init,
  n_ahead,
  n_skip = 0,
  n_lag = 0,
  mode = "slide",
  exceed = FALSE
)
}
\arguments{
\item{n_total}{The total number of observations of the time series.}

\item{n_init}{The number of periods for the initial training window (must be positive).}

\item{n_ahead}{The forecast horizon (n-steps-ahead, must be positive).}

\item{n_skip}{The number of periods to skip between windows (must be zero or positive integer).}

\item{n_lag}{A value to include a lag between the training and testing set. This is useful if lagged predictors will be used during training and testing.}

\item{mode}{Character value. Define the setup of the training window for time series cross validation. \code{stretch} is equivalent to an expanding window approach and \code{slide} is a fixed window approach.}

\item{exceed}{Logical value. If \code{TRUE}, out-of-sample splits exceeding the sample size are created.}
}
\value{
A \code{list} containing the indices for train and test as integer vectors.
}
\description{
The function creates the split indices for train and test samples
  (i.e. partitioning into time slices) for time series cross-validation. The
  user can choose between \code{stretch} and \code{slide}. The first is an
  expanding window approach, while the latter is a fixed window approach.
  The user can define the window sizes for training and testing via
  \code{n_init} and \code{n_ahead}, as well as the step size for increments
  via \code{n_step}.
}
