% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_split.R
\name{split_index}
\alias{split_index}
\title{Create indices for train and test splits.}
\usage{
split_index(
  n_total,
  n_init,
  n_ahead,
  n_skip = 0,
  n_lag = 0,
  mode = "slide",
  exceed = FALSE
)
}
\arguments{
\item{n_total}{Integer value. The total number of observations.}

\item{n_init}{Integer value. The number of periods for the initial training window (must be positive).}

\item{n_ahead}{Integer value. The forecast horizon (n-steps-ahead, must be positive).}

\item{n_skip}{Integer value. The number of periods to skip between windows (must be zero or positive integer).}

\item{n_lag}{Integer value. A value to include a lag between the training and testing set. This is useful if lagged predictors will be used during training and testing.}

\item{mode}{Character value. Define the setup of the training window for time series cross validation. \code{stretch} is equivalent to an expanding window approach and \code{slide} is a fixed window approach.}

\item{exceed}{Logical value.}
}
\value{
...
}
\description{
...
}
