% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_data.R
\name{summarise_data}
\alias{summarise_data}
\title{Summary statistics for time series data.}
\usage{
summarise_data(data)
}
\arguments{
\item{data}{A valid tsibble in long format with one measurement variable.}
}
\value{
summary_tbl A tibble containing the summary statistics.
}
\description{
This function calculates several summary statistics for time series data.
   The function calculates the following statistics for all key combinations:

   \itemize{
      \item{\code{n_missing}: Number of missing values (NAs)}
      \item{\code{complete_rate}: Percentange rate of complete values}
      \item{\code{mean}: Mean}
      \item{\code{sd}: Standard Deviation}
      \item{\code{p0}: Minimum}
      \item{\code{p25}: 25\%-Quantile}
      \item{\code{p50}: Median}
      \item{\code{p75}: 75\%-Quantile}
      \item{\code{p100}: Maximum}
      \item{\code{hist}: Histogram as sparkline}
      }
}
