% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_data.R
\name{summarise_data}
\alias{summarise_data}
\title{Summary statistics for time series data}
\usage{
summarise_data(.data, context)
}
\arguments{
\item{.data}{A \code{tibble} in long format containing time series data.}

\item{context}{A named \code{list} with the identifiers for \code{seried_id}, \code{value_id} and \code{index_id}.}
}
\value{
data A tibble containing the summary statistics.
}
\description{
This function calculates several summary statistics for time series data.
   The function calculates the following metrics for all key combinations:

   \itemize{
      \item{\code{start}: Start date (or date-time)}
      \item{\code{end}: End date (or date-time)}
      \item{\code{n_obs}: Number of observations per time series}
      \item{\code{n_missing}: Number of missing values (NAs)}
      \item{\code{pct_missing}: Percentage rate of missing values}
      \item{\code{n_zeros}: Number of zero values}
      \item{\code{pct_zeros}: Percentage rate of zero values}
      }
}
