% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_stats.R
\name{summarise_stats}
\alias{summarise_stats}
\title{Summary statistics for time series data}
\usage{
summarise_stats(.data, context)
}
\arguments{
\item{.data}{A \code{tibble} in long format containing time series data.}

\item{context}{A named \code{list} with the identifiers for \code{seried_id}, \code{value_id} and \code{index_id}.}
}
\value{
data A tibble containing the summary statistics.
}
\description{
This function calculates several summary statistics for time series data.
   The function calculates the following metrics for all key combinations:

   \itemize{
      \item{\code{mean}: Arithmetic mean}
      \item{\code{median}: Median (p50)}
      \item{\code{mode}: Mode}
      \item{\code{sd}: Standard deviation}
      \item{\code{p0}: Minimum}
      \item{\code{p25}: 25\%-Quantile}
      \item{\code{p75}: 75\%-Quantile}
      \item{\code{p100}: Maximum}
      \item{\code{skewness}: Skewness}
      \item{\code{kurtosis}: Kurtosis}
      }
}
