% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBATS.R
\name{TBATS}
\alias{TBATS}
\title{Automatic train a TBATS model}
\usage{
TBATS(formula, ...)
}
\arguments{
\item{formula}{Model specification (see "Specials" section, currently not in use ...)}

\item{...}{Further arguments passed to \code{forecast::tbats()}.}
}
\value{
tbats_model An object of class \code{TBATS}.
}
\description{
Automatic train a TBATS model (Trigonometric seasonality, Box-Cox transformation,
   ARMA errors, Trend and Seasonal components). This function is a wrapper for \code{forecast::tbats()}.
}
