% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBATS.R
\name{TBATS}
\alias{TBATS}
\title{Automatic training of TBATSs.}
\usage{
TBATS(formula, ...)
}
\arguments{
\item{formula}{Model specification (see "Specials" section, currently not in use...)}

\item{...}{Further arguments passed to \code{forecast::tbats()}.}
}
\value{
tbats_model An object of class \code{TBATS}.
}
\description{
Automatic training of TBATSs.
}
