% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSHW.R
\name{train_dshw}
\alias{train_dshw}
\title{Double Seasonal Holt-Winters model}
\usage{
train_dshw(.data, specials, periods, ...)
}
\arguments{
\item{.data}{Input data as tsibble.}

\item{specials}{Specials as list defined in \code{specials_dshw}.}

\item{periods}{Integer vector. The periodicity of the time series (e.g. \code{periods = c(24, 168)} for hourly data).}

\item{...}{Further arguments passed to \code{forecast::dshw()}.}
}
\value{
An object of class \code{DSHW}.
}
\description{
Train a Double Seasonal Holt-Winters model (DSHW).
}
