% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMEAN.R
\name{train_smean}
\alias{train_smean}
\title{Train seasonal mean model.}
\usage{
train_smean(.data, specials, period, ...)
}
\arguments{
\item{specials}{Specials as list defined in \code{specials_smean}.}

\item{period}{Integer value. The seasonal period for the calculation.}

\item{...}{Further arguments.}

\item{data}{Input data as \code{tsibble}.}
}
\value{
An object of class \code{SMEAN}.
}
\description{
Train seasonal mean model. This is equivalent to a linear regression
   against seasonal dummy variables only (\code{TSLM(value ~ season())}).
}
