% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMEDIAN.R
\name{train_smedian}
\alias{train_smedian}
\title{Seasonal median model}
\usage{
train_smedian(.data, specials, ...)
}
\arguments{
\item{.data}{Input data as tsibble.}

\item{specials}{Specials as list defined in \code{specials_smedian}.}

\item{...}{Currently not in use.}
}
\value{
An object of class \code{SMEDIAN}.
}
\description{
Train seasonal median model (SMEDIAN).
}
