% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBATS.R
\name{train_tbats}
\alias{train_tbats}
\title{Train TBATS model.}
\usage{
train_tbats(.data, specials, ...)
}
\arguments{
\item{.data}{Input data as \code{tsibble}.}

\item{specials}{Specials as list defined in \code{specials_tbats}.}

\item{...}{Further arguments passed to \code{forecast::tbats()}.}
}
\value{
An object of class \code{TBATS}.
}
\description{
Train a TBATS model (Trigonometric, Box-Cox Transform, ...).
}
