% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors_tscv.R
\name{tscv_cols}
\alias{tscv_cols}
\title{Function to extract tscv colors as hex codes.}
\usage{
tscv_cols(...)
}
\arguments{
\item{...}{Character names of tscv_colors.}
}
\description{
Function to extract tscv colors as hex codes.
}
