% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors_tscv.R
\name{tscv_pal}
\alias{tscv_pal}
\title{Return function to interpolate a tscv color palette.}
\usage{
tscv_pal(palette = "main", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character value. The name of the palette in tscv_palettes.}

\item{reverse}{Logical value indicating whether the palette should be reversed.}

\item{...}{Additional arguments to pass to colorRampPalette()}
}
\description{
Return function to interpolate a tscv color palette.
}
